% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditor.R
\name{auditor}
\alias{auditor}
\title{auditor}
\usage{
auditor(model, ...)
}
\arguments{
\item{model}{An object of appropriate class containing a model.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class ModelAudit.
}
\description{
Function \code{auditor} validates a regression model.
}
