% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous_functions.R
\name{cat_assumption}
\alias{cat_assumption}
\title{printing results of tests for chosen assumption}
\usage{
cat_assumption(ModelAudit, assumption, digits = 3)
}
\arguments{
\item{ModelAudit}{object of class ModelAudit}

\item{assumption}{name of assumption}

\item{digits}{the minimum number of significant digits to be printed in values}
}
\description{
a
}
