% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_support_function.R
\name{check_object}
\alias{check_object}
\title{Check object}
\usage{
check_object(object, type = "res")
}
\arguments{
\item{object}{Object passed to the function}
}
\description{
Checks if the object is of desired class
}
