% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_residuals.R
\name{check_residuals_autocorrelation}
\alias{check_residuals_autocorrelation}
\title{Checks for autocorrelation in target variable or in residuals}
\usage{
check_residuals_autocorrelation(model, method = "pearson")
}
\arguments{
\item{model}{an object of the class `modelAudit`}

\item{method}{will be passed to the cor.test functions}
}
\value{
autocorrelation between target variable and between residuals
}
\description{
Checks for autocorrelation in target variable or in residuals
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)
check_residuals_autocorrelation(lm_au)
}
