% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_residuals.R
\name{check_residuals_outliers}
\alias{check_residuals_outliers}
\title{Checks for outliers}
\usage{
check_residuals_outliers(model, n = 5)
}
\arguments{
\item{model}{an object of the class `modelAudit`}

\item{n}{number of lowest and highest standardized  residuals to be presented}
}
\value{
indexes of lowest and highest standardized  residuals
}
\description{
Checks for outliers
}
\examples{
library(car)
lm_model <- lm(prestige ~ education + women + income, data = Prestige)
lm_au <- audit(lm_model, data = Prestige, y = Prestige$prestige)
check_residuals_outliers(lm_au)
}
