% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_residuals.R
\name{check_residuals_trend}
\alias{check_residuals_trend}
\title{Checks for trend in residuals}
\usage{
check_residuals_trend(model, B = 20)
}
\arguments{
\item{model}{an object of the class `modelAudit`}

\item{B}{number fo samplings}
}
\value{
standardized   loess fit for residuals
}
\description{
Calculates loess fit for residuals and then extracts statistics that shows how far is this fit from one without trend
}
\examples{
library(car)
lm_model <- lm(prestige ~ education + women + income, data = Prestige)
lm_au <- audit(lm_model, data = Prestige, y = Prestige$prestige)
check_residuals_trend(lm_au)
}
