% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous_functions.R
\name{collect_assumption}
\alias{collect_assumption}
\title{Getting test name, statistic value and p.value for chosen test from ModelAudit object.}
\usage{
collect_assumption(x, name)
}
\arguments{
\item{x}{element of ModelAudit object}

\item{name}{name of assumption}
}
\description{
a
}
