% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragons.R
\docType{data}
\name{dragons}
\alias{dragons}
\title{dragons
Datasets \code{dragons} and \code{dragons_test} are artificial,
generated form the same model.

\itemize{
\item life.len - dragon's life length
\item teeth.len - dragon's teeth length 
\item colour - coulour of the dragon
\item height - dragon's height
\item herd.size - the size of the dragon herd
}}
\format{a data frame with 2000 rows and 5 columns}
\usage{
data(dragons)
}
\description{
dragons
Datasets \code{dragons} and \code{dragons_test} are artificial,
generated form the same model.

\itemize{
\item life.len - dragon's life length
\item teeth.len - dragon's teeth length 
\item colour - coulour of the dragon
\item height - dragon's height
\item herd.size - the size of the dragon herd
}
}
\keyword{dragons}
