% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragons_test.R
\docType{data}
\name{dragons_test}
\alias{dragons_test}
\title{dragons_test
Datasets \code{dragons} and \code{dragons_test} are artificial,
generated form the same model.

\itemize{
\item life.len - dragon's life length
\item teeth.len - dragon's teeth length 
\item colour - coulour of the dragon
\item height - dragon's height
\item herd.size - the size of the dragon herd
}}
\format{a data frame with 1000 rows and 5 columns}
\usage{
data(dragons_test)
}
\description{
dragons_test
Datasets \code{dragons} and \code{dragons_test} are artificial,
generated form the same model.

\itemize{
\item life.len - dragon's life length
\item teeth.len - dragon's teeth length 
\item colour - coulour of the dragon
\item height - dragon's height
\item herd.size - the size of the dragon herd
}
}
\keyword{dragons_test}
