% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_support_function.R
\name{drwhy_geom_point}
\alias{drwhy_geom_point}
\title{DrWhy's wrapper for geom_point function}
\usage{
drwhy_geom_point(df, smooth = FALSE, alpha_val)
}
\arguments{
\item{df}{Data frame prepared by (\code{make_dataframe}) function}

\item{smooth}{Logical, if set to \code{TRUE} point are drawn with alpha (set in \code{alpha_val}
argument).  Default is \code{FALSE}}

\item{alpha_val}{Numeric, level of alpha of points when smooth is drawn}
}
\description{
Function which draws point layers in desired order
}
