% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_support_function.R
\name{drwhy_geom_smooth}
\alias{drwhy_geom_smooth}
\title{DrWhy's wrapper for geom_smooth function}
\usage{
drwhy_geom_smooth(df)
}
\arguments{
\item{df}{Data frame prepared by (\code{make_dataframe}) function}
}
\description{
Function which draws smooth layers in desired order
}
