% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_support_function.R
\name{make_dataframe}
\alias{make_dataframe}
\title{Make data frame}
\usage{
make_dataframe(object, ..., variable, type = "res")
}
\arguments{
\item{object}{Objecta passed to the function}
}
\description{
Makes data frame(s) from passed models
}
