% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_support_function.R
\name{make_scale_loc_df}
\alias{make_scale_loc_df}
\title{Extra variables for scaleLocation plot}
\usage{
make_scale_loc_df(object)
}
\arguments{
\item{object}{An audited model}
}
\description{
Function to generate extra variables for scaleLocation plot
}
