% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cooksdistance.R
\name{model_cooksdistance}
\alias{model_cooksdistance}
\alias{observationInfluence}
\title{Create Observation Influence Explainer}
\usage{
model_cooksdistance(object, ...)

observationInfluence(object, ...)
}
\arguments{
\item{object}{An object of class model_audit.}

\item{...}{other parameters.}
}
\description{
Calculates Cook's distances.
}
\examples{
library(DALEX)
data(titanic)
titanic <- na.omit(titanic)
model_glm <- glm(survived ~ ., family = binomial, data = titanic)
audit_glm <- audit(model_glm)

model_cooksdistance(audit_glm)

}
