% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{model_evaluation}
\alias{model_evaluation}
\alias{modelEvaluation}
\title{Create Model Evaluation explaination}
\usage{
model_evaluation(object)

modelEvaluation(object)
}
\arguments{
\item{object}{An object of class 'explainer' created with function \code{\link[explain]{DALEX}} from the DALEX package.}
}
\value{
An object of class 'auditor_model_evaluation'.
}
\description{
Creates explanation of classification model.

Returns, among others, true positive rate (tpr), false positive rate (fpr),
rate of positive prediction (rpp), and true positives (tp).

Created object of class 'auditor_model_evaluation' can be used to plot
Receiver Operating Characteristic (ROC) curve (plot \code{\link{plot_roc}}) and LIFT curve (plot \code{\link{plot_lift}}).
}
\examples{
library(DALEX)
data(titanic)
titanic <- na.omit(titanic)
titanic$survived <- titanic$survived == "yes"
model_glm <- glm(survived ~ ., family = binomial, data = titanic)
exp_glm <- explain(model_glm, data= titanic, y = titanic$survived)

model_evaluation(exp_glm)


}
