% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{model_performance}
\alias{model_performance}
\title{Create Model Performance Explainer}
\usage{
model_performance(object, score = c("mae", "mse", "rec", "rroc"),
  new_score = NULL)
}
\arguments{
\item{object}{An object of class 'explainer' created with function \code{\link[explain]{DALEX}} from the DALEX package.}

\item{score}{Vector of score names to be plotted.}

\item{new_score}{A named list of functions that take one argument: object of class 'explainer' and return a numeric value. The measure calculated by the function should have the property that lower score value indicates better model.}

\item{...}{other parameters.}
}
\value{
An object of the class 'auditor_model_performance'.
}
\description{
Creates modelPerformance object to be plotted.
}
\examples{
library(DALEX)
data(titanic)
titanic <- na.omit(titanic)
titanic$survived <- titanic$survived == "yes"
model_glm <- glm(survived ~ ., family = binomial, data = titanic)
audit_glm <- audit(model_glm, y = titanic$survived)

model_performance(audit_glm)


}
