% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_halfnormal.R
\name{modelFit}
\alias{modelFit}
\title{Create Model Fit explainer}
\usage{
modelFit(object, quant.scale = FALSE, ...)
}
\arguments{
\item{object}{An object of class ModelAudit.}

\item{quant.scale}{if TRUE values on axis are on quantile scale.}

\item{...}{other parameters passed do \code{\link[hnp]{hnp}} function.}
}
\description{
Creates modelFit object to be plotted.
}
\examples{
library(DALEX)
data(titanic)
titanic <- na.omit(titanic[1:100,])
model_glm <- glm(survived ~ ., family = binomial, data = titanic)
audit_glm <- audit(model_glm)

modelFit(audit_glm)

}
