% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPerformance.R
\name{modelPerformance}
\alias{modelPerformance}
\title{Create Observation Influence Explainer}
\usage{
modelPerformance(object, type = c("MAE", "MSE", "REC", "RROC"),
  new.score = NULL)
}
\arguments{
\item{object}{An object of class ModelAudit.}

\item{other}{other parameters.}
}
\description{
Creates observationInfluence object to be plotted.
}
