% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observationInfluence.R
\name{observationInfluence}
\alias{observationInfluence}
\title{Create Observation Influence Explainer}
\usage{
observationInfluence(object, ...)
}
\arguments{
\item{object}{An object of class ModelAudit.}

\item{...}{other parameters.}
}
\description{
Creates observationInfluence object to be plotted.
}
\examples{
library(MASS)
model.glm <- glm(Postwt ~ Prewt + Treat + offset(Prewt), family = gaussian, data = anorexia)
audit.glm <- audit(model.glm)

oi.glm <- observationInfluence(audit.glm)

}
