% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_audit.R,
%   R/plot_model_cooksdistance.R, R/plot_model_evaluation.R,
%   R/plot_model_halfnormal.R, R/plot_model_performance.R,
%   R/plot_model_residual.R
\name{plot.model_audit}
\alias{plot.model_audit}
\alias{plot.auditor_model_cooksdistance}
\alias{plot.auditor_model_evaluation}
\alias{plot.auditor_model_halfnormal}
\alias{plot.auditor_model_performance}
\alias{plot.auditor_model_residual}
\title{Model Diagnostic Plots}
\usage{
\method{plot}{model_audit}(x, ..., type = "residual", ask = TRUE,
  grid = TRUE)

\method{plot}{auditor_model_cooksdistance}(x, ..., type = "residual",
  ask = TRUE, grid = TRUE)

\method{plot}{auditor_model_evaluation}(x, ..., type = "residual",
  ask = TRUE, grid = TRUE)

\method{plot}{auditor_model_halfnormal}(x, ..., type = "residual",
  ask = TRUE, grid = TRUE)

\method{plot}{auditor_model_performance}(x, ..., type = "residual",
  ask = TRUE, grid = TRUE)

\method{plot}{auditor_model_residual}(x, ..., type = "residual",
  ask = TRUE, grid = TRUE)
}
\arguments{
\item{x}{object of class 'auditor_model_residual' (created with \code{\link{model_residual}} function),
'auditor_model_performance' (created with \code{\link{model_performance}} function),
'auditor_model_evaluation' (created with \code{\link{model_evaluation}} function),
'auditor_model_cooksdistance' (created with \code{\link{model_cooksdistance}} function),
or 'auditor_model_halfnormal' (created with \code{\link{model_halfnormal}} function).}

\item{...}{other arguments dependent on the type of plot or additionl objects of classes 'auditor_model_residual',
'auditor_model_performance', 'auditor_model_evaluation', 'auditor_model_cooksdistance', 'auditor_model_halfnormal'.}

\item{type}{the type of plot. Character or vector of characters. Possible values: 'acf', 'autocorrelation', 'cooksdistance', 'halfnormal', 'residuals',
'lift', 'pca', 'radar', 'correlation', 'prediction', 'rec', 'resiual', 'residual_boxplot','residual_density',
'roc', 'rroc', 'scalelocation', 'tsecdf' (for detailed description see corresponding functions in see also section).}

\item{ask}{logical; if TRUE, the user is asked before each plot, see \code{\link[graphics]{par}(ask=)}.}

\item{grid}{logical; if TRUE plots will be plotted on the grid.}
}
\description{
This function provides several diagnostic plots for regression and classification models.
Provide object created with one of aduditor's computational functions, \code{\link{model_residual}},
\code{\link{model_cooksdistance}}, \code{\link{model_evaluation}}, \code{\link{model_performance}},
\code{\link{model_evaluation}}.
}
\examples{
dragons <- DALEX::dragons[1:100, ]

# fit a model
lm_model <- lm(life_length ~ ., data = dragons)

# use DALEX package to wrap up a model into explainer
lm_exp <- DALEX::explain(lm_model, data = dragons, y = dragons$life_length)

# validate a model with auditor
library(auditor)
lm_mr <- model_residual(lm_exp)

# plot results
plot(lm_mr)
plot(lm_mr, type = "prediction")

lm_hn <- model_halfnormal(lm_exp)
plot(lm_hn)

library(randomForest)
rf_model <- randomForest(life_length~., data = dragons)
rf_exp <- DALEX::explain(rf_model, data = dragons, y = dragons$life_length)

rf_mp <- model_performance(rf_exp)
lm_mp <- model_performance(lm_exp)
plot(lm_mp, rf_mp)


}
\seealso{
\code{\link{plot_acf}, \link{plot_autocorrelation}, \link{plot_cooksdistance},
\link{plot_halfnormal}, \link{plot_residual}, \link{plot_residual_boxplot}, \link{plot_lift}, \link{plot_pca},
\link{plot_radar}, \link{plot_correlation},
\link{plot_prediction}, \link{plot_rec}, \link{plot_residual_density}, \link{plot_residual}, \link{plot_roc},
\link{plot_rroc}, \link{plot_scalelocation}, \link{plot_tsecdf}}
}
