% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modelAudit.R
\name{plot.modelAudit}
\alias{plot.modelAudit}
\title{Model diagnostic plots}
\usage{
\method{plot}{modelAudit}(x, ..., type = "Residuals", ask = TRUE)
}
\arguments{
\item{x}{object of class modelAudit}

\item{...}{other arguments dependent on the type of plot or additionam objects of class modelAudit}

\item{type}{the type of plot. Possible values: 'ACF', 'Autocorrelation', 'CGains', 'Cook', 'HalfNormal', 'Residuals', 'LIFT',
ModelPCA', 'Pairs', 'Prediction', 'REC', 'ResidDens', 'Residuals', 'ROC', 'RROC',
ScaleLocation', 'TwoSidedECDF' (for detailed description see functions in seealso section).}

\item{ask}{logical; if TRUE, the user is asked before each plot, see \code{\link[graphics]{par}(ask=)}.}
}
\description{
This function provides several diagnostic plots for regression and classification models.
}
\seealso{
\code{\link{plotACF}, \link{plotAutocorrelation}, \link{plotCGains}, \link{plotCook},
\link{plotHalfNormal}, \link{plotResiduals}, \link{plotLIFT}, \link{plotModelPCA}, \link{plotPairs},
\link{plotPrediction}, \link{plotREC}, \link{plotResidDens}, \link{plotResiduals}, \link{plotROC},
\link{plotRROC}, \link{plotScaleLocation}, \link{plotTwoSidedECDF}}
}
