% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acf.R
\name{plot_acf}
\alias{plot_acf}
\title{ACF plot}
\usage{
plot_acf(object)
}
\arguments{
\item{object}{An object of class ModelAudit}
}
\description{
Plot autocorrelation function of model residuals.
Runs test, Durbin-Watson test
}
