% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_autocorr.R
\name{plot_autocorr}
\alias{plot_autocorr}
\title{Autocorrelation plot}
\usage{
plot_autocorr(object)
}
\arguments{
\item{object}{An object of class ModelAudit}
}
\description{
Plot i-th residual vs i+1-th residual. Checking autocorrelation of residuals.
Runs test, Durbin-Watson test
}
