% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_autocorrelation.R
\name{plot_autocorrelation}
\alias{plot_autocorrelation}
\alias{plotAutocorrelation}
\title{Autocorrelation of Residuals Plot}
\usage{
plot_autocorrelation(object, ..., smooth = FALSE)

plotAutocorrelation(object, ..., smooth = FALSE)
}
\arguments{
\item{object}{An object of class 'auditor_model_residual' created with \code{\link{model_residual}} function.}

\item{...}{Other 'auditor_model_residual' objects to be plotted together.}

\item{smooth}{Logical, if TRUE smooth line will be added.}
}
\value{
A ggplot object.
}
\description{
Plot of i-th residual vs i+1-th residual.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_exp <- DALEX::explain(lm_model, data = dragons, y = dragons$life_length)

library(auditor)
plot_autocorrelation(lm_exp)
plot_autocorrelation(lm_exp, score = TRUE, smooth = TRUE)

}
