% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cook_dist.R
\name{plot_cooks_dist}
\alias{plot_cooks_dist}
\title{Cooks distance plot}
\usage{
plot_cooks_dist(object, cut.off = 1)
}
\arguments{
\item{object}{An object of class ModelAudit}
}
\description{
cooks distances
}
