% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation.R
\name{plot_correlation}
\alias{plot_correlation}
\alias{plotModelCorrelation}
\title{Model Correlation Plot}
\usage{
plot_correlation(object, ..., values = "fit")

plotModelCorrelation(object, ..., values = "fit")
}
\arguments{
\item{object}{An object of class 'model_audit' or 'model_residual'.}

\item{...}{Other modelAudit or modelResiduals objects to be plotted together.}

\item{values}{"fit" for model fitted values or "res" for residual values.}
}
\value{
Invisibly returns a \code{\link[gtable]{gtable}} object.
}
\description{
Matrix of plots
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)

library(randomForest)
rf_model <- randomForest(life_length~., data = dragons)
rf_au <- audit(rf_model, data = dragons, y = dragons$life_length)
plotModelCorrelation(lm_au, rf_au)

}
\seealso{
\code{\link{plot.model_audit}}
}
