% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_halfnormal.R
\name{plot_halfnormal}
\alias{plot_halfnormal}
\alias{plotHalfNormal}
\title{Half-Normal plot}
\usage{
plot_halfnormal(object, ..., quantiles = FALSE, sim = 99)

plotHalfNormal(object, ..., quantiles = FALSE, sim = 99)
}
\arguments{
\item{object}{'model_audit' object, 'model_halfnormal' object.}

\item{...}{extra arguments passed to \link[hnp]{hnp}.}

\item{quantiles}{if TRUE values on axis are on quantile scale.}

\item{sim}{number of residuals to simulate.}
}
\value{
An object of class ggplot
}
\description{
The half-normal plot is one of the tools designed to evaluate the goodness of fit of a
statistical models. It is a graphical method for comparing two probability distributions by plotting
their quantiles against each other.
Points on the plot correspond to ordered absolute values of model diagnostic
(i.e. standardized residuals) plotted against theoretical order statistics from a half-normal distribution.
}
\seealso{
\code{\link{score_halfnormal}}
}
