% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residual.R
\name{plot_residual}
\alias{plot_residual}
\alias{plotResidual}
\title{Plot Residuals vs Observed, Fitted or Variable Values}
\usage{
plot_residual(object, ..., variable = "_y_", smooth = FALSE,
  std_residuals = FALSE, nlabel = 0)

plotResidual(object, ..., variable = NULL, smooth = FALSE,
  std_residuals = FALSE, nlabel = 0)
}
\arguments{
\item{object}{An object of class 'model_audit' or 'model_residual'.}

\item{...}{Other modelAudit objects to be plotted together.}

\item{variable}{Name of variable to order residuals on a plot.
If \code{variable="_y_"}, the data is ordered by a vector of actual response (\code{y} parameter
passed to the \code{\link{explain}} function).
If \code{variable = "_y_hat_"} the data on the plot will be ordered by predicted response.
If \code{variable = NULL}, unordered observations are presented.}

\item{smooth}{Logical, indicates whenever smoothed lines should be added. By default it's FALSE.}

\item{std_residuals}{Logical, indicates whenever standardized residuals should be used.}

\item{nlabel}{Number of observations with the biggest Cook's distances to be labeled.}
}
\description{
A plot of residuals against fitted values, observed values or any variable.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_exp <- DALEX::explain(lm_model, data = dragons, y = dragons$life_length)
lm_mr <- model_residual(lm_exp)
plot_residual(lm_mr)

library(randomForest)
rf_model <- randomForest(life_length~., data = dragons)
rf_au <- audit(rf_model, data = dragons, y = dragons$life_length)
plot_residual(lm_au, rf_au)

}
\seealso{
\code{\link{plot.model_audit}}
}
