% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scalelocation.R
\name{plot_scalelocation}
\alias{plot_scalelocation}
\alias{plotScaleLocation}
\title{Scale location plot}
\usage{
plot_scalelocation(object, ..., variable = "_y_", smooth = FALSE,
  peaks = FALSE)

plotScaleLocation(object, ..., variable = NULL, smooth = FALSE,
  peaks = FALSE)
}
\arguments{
\item{object}{An object of class 'model_audit' or 'model_residual'.}

\item{...}{Other modelAudit objects to be plotted together.}

\item{variable}{Name of variable to order residuals on a plot.
If \code{variable="_y_"}, the data is ordered by a vector of actual response (\code{y} parameter
passed to the \code{\link{explain}} function).
If \code{variable = "_y_hat_"} the data on the plot will be ordered by predicted response.
If \code{variable = NULL}, unordered observations are presented.}

\item{smooth}{Logical, indicates whenever smoothed lines should be added. By default it's FALSE.}

\item{peaks}{A logical value. If TRUE peaks are marked on plot by black dots.}
}
\description{
Variable values vs square root of the absolute value of the residuals.
A vertical line corresponds to median.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_exp <- DALEX::explain(lm_model, data = dragons, y = dragons$life_length)
library(auditor)
lm_mr <- model_residual(lm_exp)
plot_scalelocation(lm_mr)


}
