% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_test_gq.R
\name{plot_test_gq}
\alias{plot_test_gq}
\title{Scale location plot}
\usage{
plot_test_gq(object)
}
\arguments{
\item{object}{An object of class ModelAudit}
}
\description{
Variable values vs square root of the absolute value of the residuals.
A vertical line corresponds to median.
Goldfeld-Quandt test - checking assumption of homoscedasticity of residuals.
}
