% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vif.R
\name{plot_vif}
\alias{plot_vif}
\title{VIF plot}
\usage{
plot_vif(object, cut.off = 10)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{cut.off}{horizontal line}
}
\description{
VIF (Variance inflation factor) computed for each variable
}
