% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotACF.R
\name{plotACF}
\alias{plotACF}
\title{ACF plot}
\usage{
plotACF(object, variable = NULL)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{variable}{name of dependent or independent variable to order residuals. If NULL the fitted values are taken.}
}
\description{
Plot autocorrelation function of model residuals.
Runs test, Durbin-Watson test
}
