% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAutocorr.R
\name{plotAutocorr}
\alias{plotAutocorr}
\title{Autocorrelation plot}
\usage{
plotAutocorr(object, variable)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{variable}{name of variable to order residuals}
}
\description{
Plot i-th residual vs i+1-th residual. Checking autocorrelation of residuals.
Runs test, Durbin-Watson test
}
