% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAutocorrelation.R
\name{plotAutocorrelation}
\alias{plotAutocorrelation}
\title{Autocorrelation Plot}
\usage{
plotAutocorrelation(object, ..., variable = NULL, smooth = FALSE)
}
\arguments{
\item{object}{An object of class modelAudit or modelResiduals.}

\item{...}{Other modelAudit or modelResiduals objects to be plotted together.}

\item{variable}{Only for modelAudit objects. Name of model variable to order residuals.
If value is NULL the data is ordered by a vector of actual response (\code{y} parameter
passed to the \code{\link{audit}} function). One can also pass any name of any other variable
in the data set. If \code{variable = ""} is set, unordered observations are presented.}

\item{smooth}{Logical, if TRUE smooth line will be added.}
}
\description{
Plot of i-th residual vs i+1-th residual.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)
plotAutocorrelation(lm_au)
plotAutocorrelation(lm_au, variable = "height", score = TRUE, smooth = TRUE)

}
