% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCGains.R
\name{plotCGains}
\alias{plotCGains}
\title{Cumulative Gains Chart}
\usage{
plotCGains(object, ..., newdata = NULL, newy)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{...}{other modelAudit objects to be plotted together}

\item{newdata}{optionally, a data frame in which to look for variables with which to plot CGains curve. If omitted, the data used to build model will be used.}

\item{newy}{optionally, required if newdata used. Response vector for new data.}
}
\value{
ggplot object
}
\description{
Cumulative Gains Chartis a plot of the rate of positive prediction against true positive rate for the different thresholds.
It is useful for measuring and comparing the accuracy of the classificators.
}
\examples{
library(auditor)
library(mlbench)
library(randomForest)
data("PimaIndiansDiabetes")

model_rf <- randomForest(diabetes~., data=PimaIndiansDiabetes)
au_rf <- audit(model_rf, label="rf")
plotCGains(au_rf)

model_glm <- glm(diabetes~., family=binomial,	data=PimaIndiansDiabetes)
au_glm <- audit(model_glm)
plotCGains(au_rf, au_glm)

}
\seealso{
\code{\link{plot.modelAudit}}
}
