% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCook.R
\name{plotCook}
\alias{plotCook}
\title{Cook plot}
\usage{
plotCook(object, nlabel = 3)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{nlabel}{number of observations with the biggest Cooks distances to be labeled}
}
\description{
Cooks distances
}
