% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCumulativeGain.R
\name{plotCumulativeGain}
\alias{plotCumulativeGain}
\title{Cumulative Gain Chart}
\usage{
plotCumulativeGain(object, ...)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{...}{other modelAudit objects to be plotted together}
}
\value{
ggplot object
}
\description{
Cumulative Gain Chartis is a plot of the rate of positive prediction against true positive rate for the different thresholds.
It is useful for measuring and comparing the accuracy of the classificators.
}
\seealso{
\code{\link{plot.modelAudit}}
}
