% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotD3_cooksdistance.R
\name{plotD3_cooksdistance}
\alias{plotD3_cooksdistance}
\alias{plotD3CooksDistance}
\title{Influence of observations Plot in D3 with r2d3 package.}
\usage{
plotD3_cooksdistance(object, ..., nlabel = 3, single_plot = FALSE,
  scale_plot = FALSE, background = FALSE)

plotD3CooksDistance(object, ..., nlabel = 3, single_plot = FALSE,
  scale_plot = FALSE, background = FALSE)
}
\arguments{
\item{object}{An object of class 'model_audit' or 'model_cooksdistance'.}

\item{...}{Other modelAudit or modelResiduals objects to be plotted together.}

\item{nlabel}{Number of observations with the biggest Cook's distances to be labeled.}

\item{single_plot}{Logical, indicates whenever single or facets should be plotted. By default it's FALSE.}

\item{scale_plot}{Logical, indicates whenever the plot should scale with height. By default it's FALSE.}

\item{background}{Logical, available only if single_plot = FALSE. Indicates whenever backgroud plots should be plotted. By default it's FALSE.}
}
\value{
a `r2d3` object.
}
\description{
Cook’s distances are used for estimate the influence of an single observation.
If the picture is not displayed in the viewer, please update your RStudio.
}
\seealso{
\code{\link{plot_cooksdistance}}
}
