% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotD3_lift.R
\name{plotD3_lift}
\alias{plotD3_lift}
\alias{plotD3LIFT}
\title{Plot LIFT in D3 with r2d3 package.}
\usage{
plotD3_lift(object, ..., scale_plot = FALSE)

plotD3LIFT(object, ..., scale_plot = FALSE)
}
\arguments{
\item{object}{An object of class 'model_audit' or 'model_evaluation'.}

\item{...}{Other modelAudit objects to be plotted together.}

\item{scale_plot}{Logical, indicates whenever the plot should scale with height. By default it's FALSE.}
}
\value{
a `r2d3` object
}
\description{
LIFT is a plot of the rate of positive prediction against true positive rate for the different thresholds.
It is useful for measuring and comparing the accuracy of the classificators.
}
\seealso{
\code{\link{plot_lift}}
}
