% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotD3CooksDistance.R
\name{plotD3CooksDistance}
\alias{plotD3CooksDistance}
\title{Influence of observations Plot in D3 with r2d3 Package.}
\usage{
plotD3CooksDistance(object, ..., nlabel = 3, single_plot = FALSE,
  scale_plot = FALSE, background = FALSE)
}
\arguments{
\item{object}{An object of class modelAudit or observationInfluence.}

\item{...}{Other modelAudit or modelResiduals objects to be plotted together.}

\item{nlabel}{Number of observations with the biggest Cook's distances to be labeled.}

\item{single_plot}{Logical, indicates whenever single or facets should be plotted. By default it's FALSE.}

\item{scale_plot}{Logical, indicates whenever the plot should scale with height. By default it's FALSE.}

\item{background}{Logical, available only if single_plot = FALSE. Indicates whenever backgroud plots should be plotted. By default it's FALSE.}
}
\description{
Cook’s distances are used for estimate the influence of an single observation.
If the picture is not displayed in the viewer, please update your RStudio.
}
\examples{
library("auditor")
library(car)

lm_model <- lm(prestige~education + women + income, data = Prestige)
lm_au <- audit(lm_model, data = Prestige, y = Prestige$prestige, label = "lm")
plotD3CooksDistance(lm_au)

library(randomForest)
rf_model <- randomForest(prestige~education + women + income, data = Prestige)
rf_au <- audit(rf_model, data = Prestige, y = Prestige$prestige, label = "rf")
plotD3CooksDistance(lm_au, rf_au, nlabel = 4)

}
\seealso{
\code{\link{plotCooksDistance}}
}
