% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotD3HalfNormal.R
\name{plotD3HalfNormal}
\alias{plotD3HalfNormal}
\title{Plot Half-Normal in D3 with r2d3 package.}
\usage{
plotD3HalfNormal(object, ..., quantiles = FALSE, sim = 99,
  scale_plot = FALSE)
}
\arguments{
\item{object}{modelAudit object, modelFit object.}

\item{...}{extra arguments passed to \link[hnp]{hnp}.}

\item{quantiles}{if TRUE values on axis are on quantile scale.}

\item{sim}{number of residuals to simulate}

\item{scale_plot}{Logical, indicates whenever the plot should scale with height. By default it's FALSE.}
}
\value{
a `r2d3` object.
}
\description{
The half-normal plot is one of the tools designed to evaluate the goodness of fit of a
statistical models. It is a graphical method for comparing two probability distributions by plotting
their quantiles against each other.
Points on the plot correspond to ordered absolute values of model diagnostic
(i.e. standardized residuals) plotted against theoretical order statistics from a half-normal distribution.
}
\seealso{
\code{\link{scoreHalfNormal}, \link{plotHalfNormal}}
}
