% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotD3LIFT.R
\name{plotD3LIFT}
\alias{plotD3LIFT}
\title{Plot LIFT in D3 with r2d3 Package.}
\usage{
plotD3LIFT(object, ..., scale_plot = FALSE)
}
\arguments{
\item{object}{An object of class modelAudit or modelEvaluation.}

\item{...}{Other modelAudit objects to be plotted together.}

\item{scale_plot}{Logical, indicates whenever the plot should scale with height. By default it's FALSE.}
}
\value{
a `r2d3` object
}
\description{
LIFT is a plot of the rate of positive prediction against true positive rate for the different thresholds.
It is useful for measuring and comparing the accuracy of the classificators.
}
\examples{

library(mlbench)
library("auditor")
data("PimaIndiansDiabetes")
Pima <- PimaIndiansDiabetes
Pima$diabetes <- ifelse(Pima$diabetes == "pos", 1, 0)

glm_model <- glm(diabetes ~ ., family = binomial, data=Pima)
glm_au <- audit(glm_model, data = Pima, y = Pima$diabetes, label = "glm1")

plotD3LIFT(glm_au)

glm_model2 <- glm(diabetes~pressure, family=binomial,	data=PimaIndiansDiabetes)
glm_au2 <- audit(glm_model2, data = Pima, y = Pima$diabetes, label = "glm2")

plotD3LIFT(glm_au, glm_au2, scale_plot = TRUE)

}
\seealso{
\code{\link{plotLIFT}}
}
