% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotD3_residual.R
\name{plotD3Residual}
\alias{plotD3Residual}
\title{Plot Residuals vs Target, Observed or Variable Values in D3 with r2d3 package.}
\usage{
plotD3Residual(object, ..., variable = NULL, points = TRUE,
  smooth = FALSE, std_residuals = FALSE, point_count = NULL,
  single_plot = TRUE, scale_plot = FALSE, background = FALSE)
}
\arguments{
\item{object}{An object of class modelAudit or modelResiduals.}

\item{...}{Other modelAudit or modelResiduals objects to be plotted together.}

\item{variable}{Only for modelAudit objects. Name of model variable to order residuals.
If value is NULL the data is ordered by a vector of actual response (\code{y} parameter
passed to the \code{\link{audit}} function). One can also pass any name of any other variable
in the data set. If \code{variable = ""} is set, unordered observations are presented.}

\item{points}{Logical, indicates whenever observations should be added as points. By defaul it's TRUE.}

\item{smooth}{Logical, indicates whenever smoothed lines should be added. By default it's FALSE.}

\item{std_residuals}{Logical, indicates whenever standardized residuals should be used. By default it's FALSE.}

\item{point_count}{Number of points to be plotted per model. Points will be chosen randomly. By default plot all of them.}

\item{single_plot}{Logical, indicates whenever single or facets should be plotted. By default it's TRUE.}

\item{scale_plot}{Logical, indicates whenever the plot should scale with height. By default it's FALSE.}

\item{background}{Logical, available only if single_plot = FALSE. Indicates whenever backgroud plots should be plotted. By default it's FALSE.}
}
\value{
a `r2d3` object.
}
\description{
Function \code{plotD3Residual} plots resudial values vs target, observed or variable values in the model.
It uses output from \code{modelAudit} or \code{modelResiduals} function.

If the picture is not displayed in the viewer, please update your RStudio.
}
\examples{
library("auditor")
library("DALEX")

lm_model <- lm(m2.price ~., data = apartments)
lm_au <- audit(lm_model, label = "lm")
plotD3Residual(lm_au, variable = "construction.year")

library(randomForest)
rf_model <- randomForest(m2.price ~., data = apartments)
rf_au <- audit(rf_model, label = "rf")
rf_mr <- modelResiduals(rf_au, "construction.year")
plotD3Residual(lm_au, rf_mr, variable = "construction.year", smooth = TRUE)
plotD3Residual(lm_au, rf_mr, variable = "construction.year", smooth = TRUE, single_plot = FALSE)

}
\seealso{
\code{\link{plotResidual}}
}
