% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHalfNormal.R
\name{plotHalfNormal}
\alias{plotHalfNormal}
\title{Half-Normal plot}
\usage{
plotHalfNormal(object, quant = FALSE, ...)
}
\arguments{
\item{object}{modelAudit object, modelFit object.}

\item{quant}{if TRUE values on axis are on quantile scale.}

\item{...}{extra arguments passed to \link[hnp]{hnp}.}
}
\value{
An object of class ggplot
}
\description{
The half-normal plot is one of the tools designed to evaluate the goodness of fit of a
statistical models. It is a graphical method for comparing two probability distributions by plotting
their quantiles against each other.
Points on the plot correspond to ordered absolute values of model diagnostic
(i.e. standardized residuals) plotted against theoretical order statistics from a half-normal distribution.
}
\seealso{
\code{\link{scoreHalfNormal}}
}
