% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLift.R
\name{plotLIFT}
\alias{plotLIFT}
\title{Lift Chart}
\usage{
plotLIFT(object, ..., newdata = NULL, newy, groups = 10,
  cumulative = TRUE)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{...}{other modelAudit objects to be plotted together}

\item{newdata}{optionally, a data frame in which to look for variables with which to plot CGains curve. If omitted, the data used to build model will be used.}

\item{newy}{optionally, required if newdata used. Response vector for new data.}

\item{groups}{number of groups}

\item{cumulative}{boolean. If TRUE cumulative lift curve will be plotted.}
}
\value{
ggplot object
}
\description{
Lift Chart shows the ratio of a model to a random guess.
}
\examples{
library(auditor)
library(mlbench)
library(randomForest)
data("PimaIndiansDiabetes")

model_rf <- randomForest(diabetes~., data=PimaIndiansDiabetes)
au_rf <- audit(model_rf, label="rf")
plotLIFT(au_rf)

model_glm <- glm(diabetes~., family=binomial,	data=PimaIndiansDiabetes)
au_glm <- audit(model_glm)
plotLIFT(au_rf, au_glm)

}
\seealso{
\code{\link{plot.modelAudit}}
}
