% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelPCA.R
\name{plotModelPCA}
\alias{plotModelPCA}
\title{Model PCA plot}
\usage{
plotModelPCA(object, ...)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{...}{other modelAudit objects to be plotted together}
}
\value{
ggplot object
}
\description{
PCA for model residuals
}
\seealso{
\code{\link{plot.modelAudit}}
}
