% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrediction.R
\name{plotPrediction}
\alias{plotPrediction}
\title{Predicted response vs Observed or Variable Values}
\usage{
plotPrediction(object, ..., variable = "Observed response")
}
\arguments{
\item{object}{An object of class modelAudit.}

\item{...}{Other modelAudit objects to be plotted together.}

\item{variable}{Name of model variable to order residuals. If value is NULL data order is taken. If value is "Observed response" the data is ordered by a vector of actual response (\code{y} parameter passed to the \code{\link{audit}} function).}
}
\description{
Plot of predicted response vs observed or variable Values.
}
\seealso{
\code{\link{plot.modelAudit}}
}
