% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResiduals.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Residuals plot}
\usage{
plotResiduals(object, variable = NULL)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{variable}{name of dependent or independent variable to order residuals. If NULL the fitted values are taken.}
}
\description{
Residuals vs fitted values or dependent or independent variable
}
