% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plotROC}
\alias{plotROC}
\title{Receiver Operating Characteristic (ROC)}
\usage{
plotROC(object, ..., nlabels = NULL)
}
\arguments{
\item{object}{An object of class ModelAudit or modelEvaluation.}

\item{...}{Other modelAudit objects to be plotted together.}

\item{nlabels}{Number of cutoff points to show on the plot. Default is `NULL`.}
}
\value{
ggplot object
}
\description{
Receiver Operating Characterstic Curve is a plot of the true positive rate (TPR) against the false positive rate (FPR) for the different thresholds.
It is useful for measuring and comparing the accuracy of the classificators.
}
\examples{
library(DALEX)
data(titanic)
titanic <- na.omit(titanic)
titanic$survived <- titanic$survived == "yes"
model_glm <- glm(survived ~ ., family = binomial, data = titanic)
audit_glm <- audit(model_glm, y = titanic$survived)

plotROC(audit_glm)

}
\seealso{
\code{\link{plot.model_audit}, \link{plotRROC}, \link{plotREC}}
}
