% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRROC.R
\name{plotRROC}
\alias{plotRROC}
\title{Regression Receiver Operating Characteristic (RROC)}
\usage{
plotRROC(object, ...)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{...}{other modelAudit objects to be plotted together}
}
\value{
ggplot object
}
\description{
The basic idea of the ROC curves for regression is to show model asymmetry.
The RROC is a plot where on the x-axis we depict total over-estimation and on the y-axis total
under-estimation.
}
\details{
For RROC curves we use a shift, which is an equvalent to the threshold for ROC curves.
For each observation we calculate new prediction: \eqn{\hat{y}'=\hat{y}+s}
 where s is the shift.
\eqn{OVER= \sum(e_i|e_i>0)}
 \eqn{UNDER = \sum(e_i|e_i<0)}
 The shift equals 0 is represented by dot.
}
\references{
Hernández-Orallo, J. (2013). ROC curves for regression. Pattern Recognition, 46, 3395-3411.
}
