% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTestGQ.R
\name{plotTestGQ}
\alias{plotTestGQ}
\title{Scale location plot}
\usage{
plotTestGQ(object, variable)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{variable}{name of variable to order residuals}
}
\description{
Variable values vs square root of the absolute value of the residuals.
A vertical line corresponds to median.
Goldfeld-Quandt test - checking assumption of homoscedasticity of residuals.
}
