% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.auditor_score}
\alias{print.auditor_score}
\title{Prints of Models Scores}
\usage{
\method{print}{auditor_score}(x, ...)
}
\arguments{
\item{x}{an object \code{auditor_score} created with \code{\link{score}} function.}

\item{...}{other parameters}
}
\description{
Prints of Models Scores
}
\examples{
library(DALEX)

# fit a model
model_glm <- glm(survived ~ ., family = binomial, data = titanic_imputed)
# create an explainer
exp_glm <- explain(model_glm,
                   data = titanic_imputed,
                   y = titanic_imputed$survived)
# calculate score
score(exp_glm, type = "auc")

}
