% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_functions.R
\name{print.modelAudit}
\alias{print.modelAudit}
\title{Print ModelAudit object}
\usage{
\method{print}{modelAudit}(x, ...)
}
\arguments{
\item{x}{an objecy of class ModelAudit.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{print.ModelAudit} prints its argument.
}
