% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_acc.R
\name{score_acc}
\alias{score_acc}
\title{Accuracy}
\usage{
score_acc(object, data = NULL)
}
\arguments{
\item{object}{An object of class \code{explainer} created with function \code{\link[DALEX]{explain}} from the DALEX package.}

\item{data}{New data that will be used to calcuate the score. Pass \code{NULL} if you want to use \code{data} from \code{object}.}
}
\value{
An object of class \code{auditor_score}.
}
\description{
Accuracy
}
\examples{
titanic <- na.omit(DALEX::titanic)
titanic$survived <- titanic$survived == "yes"

# fit a model
model_glm <- glm(survived ~ ., family = binomial, data = titanic)

#create an explainer
exp_glm <- DALEX::explain(model_glm, y = titanic$survived)

# calculate score
score_acc(exp_glm)


}
