% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_gini.R
\name{score_gini}
\alias{score_gini}
\title{Gini Coefficient}
\usage{
score_gini(object, data = NULL)
}
\arguments{
\item{object}{An object of class \code{explainer} created with function \code{\link[DALEX]{explain}} from the DALEX package.}

\item{data}{New data that will be used to calcuate the score. Pass \code{NULL} if you want to use \code{data} from \code{object}.}
}
\value{
An object of class \code{auditor_score}.
}
\description{
The Gini coefficient measures the inequality among values of a frequency distribution.
A Gini coefficient equals 0 means perfect equality, where all values are the same.
A Gini coefficient equals 100% expresses maximal inequality of values.
}
\examples{
titanic <- na.omit(DALEX::titanic)
titanic$survived <- titanic$survived == "yes"

# fit a model
model_glm <- glm(survived ~ ., family = binomial, data = titanic)

#create an explainer
exp_glm <- DALEX::explain(model_glm, y = titanic$survived)

# calculate score
score_gini(exp_glm)

}
\seealso{
\code{\link{plot_roc}}
}
