% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_mae.R
\name{score_mae}
\alias{score_mae}
\alias{scoreMAE}
\title{Mean Absolute Error}
\usage{
score_mae(object)

scoreMAE(object)
}
\arguments{
\item{object}{An object of class 'Model_audit' or 'model_residual'.}
}
\value{
an object of class 'score_audit'
}
\description{
Mean Absolute Error.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)
score_mae(lm_au)


}
\seealso{
\code{\link{score}}
}
