% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_roc.R, R/score_rroc.R
\name{score_roc}
\alias{score_roc}
\alias{scoreROC}
\alias{scoreRROC}
\title{Area Under ROC Curve (AUC)}
\usage{
score_roc(object)

scoreROC(object)

scoreRROC(object)
}
\arguments{
\item{object}{An object of class 'model_audit'.}
}
\value{
an object of class scoreAudit
}
\description{
Area Under Curve (AUC) for Receiver Operating Characteristic.
}
\examples{
library(DALEX)
data(titanic)
titanic <- na.omit(titanic)
titanic$survived <- titanic$survived == "yes"
model_glm <- glm(survived ~ ., family = binomial, data = titanic)
audit_glm <- audit(model_glm, y = titanic$survived)

score_roc(audit_glm)

}
\seealso{
\code{\link{plot_roc}}
}
