% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_specificity.R
\name{score_specificity}
\alias{score_specificity}
\title{Specificity}
\usage{
score_specificity(object)
}
\arguments{
\item{object}{An object of class 'explainer' created with function \code{\link[DALEX]{explain}} from the DALEX package.}
}
\value{
An object of class 'auditor_score'.
}
\description{
Specificity
}
\examples{
titanic <- na.omit(DALEX::titanic)
titanic$survived <- titanic$survived == "yes"

# fit a model
model_glm <- glm(survived ~ ., family = binomial, data = titanic)

#create an explainer
exp_glm <- DALEX::explain(model_glm, y = titanic$survived)

# calculate score
score_specificity(exp_glm)


}
