% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreDW.R
\name{scoreDW}
\alias{scoreDW}
\title{Durbin-Watson Score}
\usage{
scoreDW(object, variable = NULL)
}
\arguments{
\item{object}{object An object of class ModelAudit}

\item{variable}{"Fitted values" or name of dependent or independent variable to order residuals. If NULL the original data order is taken.}
}
\value{
an object of class scoreAudit
}
\description{
Score based on Durbin-Watson test statistic.
The score value is helpful in comparing models. It is worth ponting out that results of tests like p-value makes sense only
when the test assumptions are satisfied. Otherwise test statistic may be considered as a score.
}
