% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreHalfNormal.R
\name{scoreHalfNormal}
\alias{scoreHalfNormal}
\title{Half-Normal Score}
\usage{
scoreHalfNormal(object, ...)
}
\arguments{
\item{object}{ModelAudit object or fitted model}

\item{...}{extra arguments passed to \link[hnp]{hnp}.}
}
\description{
Simulated score
}
