% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMAE.R
\name{scoreMAE}
\alias{scoreMAE}
\title{Mean Absolute Error}
\usage{
scoreMAE(object)
}
\arguments{
\item{object}{An object of class ModelAudit.}
}
\value{
an object of class scoreAudit
}
\description{
Mean Absolute Error.
}
\seealso{
\code{\link{score}}
}
