% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMSE.R
\name{scoreMSE}
\alias{scoreMSE}
\title{Mean Square Error}
\usage{
scoreMSE(object)
}
\arguments{
\item{object}{An object of class ModelAudit.}
}
\value{
an object of class scoreAudit
}
\description{
Mean Square Error.
}
\examples{
library(car)
lm_model <- lm(prestige~education + women + income, data = Prestige)
lm_au <- audit(lm_model, data = Prestige, y = Prestige$prestige)
scoreMSE(lm_au)

}
\seealso{
\code{\link{score}}
}
