% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreREC.R
\name{scoreREC}
\alias{scoreREC}
\title{Area Over the Curve for REC Curves}
\usage{
scoreREC(object)
}
\arguments{
\item{object}{An object of class ModelAudit}
}
\value{
numeric
}
\description{
The area over the Regression Error Characteristic curve is a measure of the expected error
for the regression model.
It is calculated as a sample mean of errors.
}
\seealso{
\code{\link{plotREC}}
}
