% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreRMSE.R
\name{scoreRMSE}
\alias{scoreRMSE}
\title{Root Mean Square Error}
\usage{
scoreRMSE(object)
}
\arguments{
\item{object}{An object of class ModelAudit.}
}
\value{
an object of class scoreAudit
}
\description{
Root Mean Square Error.
}
\examples{
library(car)
lm_model <- lm(prestige~education + women + income, data = Prestige)
lm_au <- audit(lm_model, data = Prestige, y = Prestige$prestige)
scoreRMSE(lm_au)


}
\seealso{
\code{\link{score}}
}
