% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreRuns.R
\name{scoreRuns}
\alias{scoreRuns}
\title{Runs Score}
\usage{
scoreRuns(object, variable = NULL)
}
\arguments{
\item{object}{object An object of class ModelAudit}

\item{variable}{"Fitted values" or name of dependent or independent variable to order residuals. If NULL original data order is taken.}
}
\description{
Score based on Runs test statistic. Note that this test is not very strong. It utilizes only signs of the residuals.
The score value is helpful in comparing models. It is worth ponting out that results of tests like p-value makes sense only
when the test assumptions are satisfied. Otherwise test statistic may be considered as a score.
}
