% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_functions.R
\name{summary.ModelAudit}
\alias{summary.ModelAudit}
\title{summary of a ModelAudit object}
\usage{
\method{summary}{ModelAudit}(object, ...)
}
\arguments{
\item{object}{an object of class ModelAudit.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Function \code{summary.ModelAudit} summarizes object of class ModelAudit.
}
