% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dw.R
\name{test_dw}
\alias{test_dw}
\title{Durbin-Watson Test}
\usage{
test_dw(object, variable = NULL)
}
\arguments{
\item{object}{model or ordered vector of residuals}

\item{variable}{name of variable - the observations in the model are ordered by the values of this variable. If NULL - first variable is taken.}
}
\description{
Test for autocorrelation with the Durbin-Watson Test.
}
