% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_gq.R
\name{test_gq}
\alias{test_gq}
\title{Goldfeld-Quandt Test}
\usage{
test_gq(model, variable = NULL)
}
\arguments{
\item{model}{model}

\item{variable}{name of variable - the observations in the model are ordered by the values of this variable. If NULL - first variable is taken.}
}
\description{
Test for homooskedasticity with the Goldfeld-Quandt Test.
}
