% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_runs.R
\name{test_runs}
\alias{test_runs}
\title{Runs Test}
\usage{
test_runs(object, variable = NULL)
}
\arguments{
\item{object}{model or ordered vector of residuals}

\item{variable}{name of variable - the observations in the model are ordered by the values of this variable. If NULL - first variable is taken.}
}
\description{
Note that this test is not very strong. It utilizes only signs of the residuals.
}
