% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testRuns.R
\name{testRuns}
\alias{testRuns}
\title{Runs Test}
\usage{
testRuns(model, vars = NULL)
}
\arguments{
\item{model}{model}

\item{vars}{name of variable - the observations in the model are ordered by the values of this variable. If NULL - first variable is taken.}
}
\description{
Note that this test is not very strong. It utilizes only signs of the residuals.
}
