% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{audit}
\alias{audit}
\title{audit}
\usage{
audit(model)
}
\arguments{
\item{model}{An object of appropriate class containing a model}
}
\value{
An object of class ModelAudit.
}
\description{
Function \code{audit} create ModelAudit object for further validation of model.
}
