% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit.R
\name{audit}
\alias{audit}
\title{audit}
\usage{
audit(model, data = NULL, y = NULL)
}
\arguments{
\item{model}{An object of appropriate class containing a model}

\item{data}{data.frame or matrix - data used for fitting. If not provided, will be extracted from the model.}

\item{y}{response used for building a model}
}
\value{
An object of class ModelAudit.
}
\description{
Function \code{audit} create ModelAudit object for further validation of model.
}
