% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditor.R
\name{auditor}
\alias{auditor}
\title{auditor}
\usage{
auditor(model, variable = NULL)
}
\arguments{
\item{model}{An object of appropriate class containing a model.}

\item{variable}{name of variable - for some tests the observations will be ordered by the values of this variable. If NULL - first variable is taken.}
}
\value{
An object of class ModelAudit.
}
\description{
Function \code{auditor} validates a regression model.
}
