% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditor.R
\name{auditor}
\alias{auditor}
\title{auditor}
\usage{
auditor(model, vars = NULL)
}
\arguments{
\item{model}{An object of appropriate class containing a model.}

\item{vars}{vector of variables - for some tests and plots observations in the model are ordered by the values of this variables. If NULL - all variables are taken.}
}
\value{
An object of class ModelAudit.
}
\description{
Function \code{auditor} validates a regression model.
}
