% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditorData.R
\docType{data}
\name{auditorData}
\alias{auditorData}
\title{Artificial auditorData}
\format{
a data frame with 2000 rows and 5 columns
}
\usage{
data(auditorData)
}
\description{
The auditor Data is an artificial data set.
It consists of 2000 observations.
First four of simulated variables are treated as continuous while the fifth one is categorical.
}
\examples{
data("auditorData", package = "auditor")
head(auditorData)
}
