% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_residuals.R
\name{check_residuals_outliers}
\alias{check_residuals_outliers}
\title{Checks for outliers}
\usage{
check_residuals_outliers(model, n = 5)
}
\arguments{
\item{model}{an object of the class `modelAudit`}

\item{n}{number of lowest and highest standardized  residuals to be presented}
}
\value{
indexes of lowest and highest standardized  residuals
}
\description{
Checks for outliers
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)
check_residuals_outliers(lm_au)
}
