% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_support_function.R
\name{make_dataframe}
\alias{make_dataframe}
\title{Make data frame}
\usage{
make_dataframe(object, ..., variable = NULL, nlabel = NULL,
  type = "res")
}
\arguments{
\item{object}{Object passed to the function}

\item{...}{Other modelAudit objects to be plotted together}

\item{variable}{Variable}

\item{type}{Type of check; default is \code{res} which stands for "model residuals".
Other possible values: \code{eva} - model evaluation}
}
\description{
Makes data frame(s) from passed models
}
