% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_halfnormal.R
\name{model_halfnormal}
\alias{model_halfnormal}
\title{Create Halfnormal Explanation}
\usage{
model_halfnormal(object, quant = FALSE, ...)
}
\arguments{
\item{object}{An object of class 'explainer' created with function \code{\link[explain]{DALEX}} from the DALEX package.}

\item{quant}{if TRUE values on axis are on quantile scale.}

\item{...}{other parameters passed do \code{\link[hnp]{hnp}} function.}
}
\value{
An object of the class 'auditor_model_halfnormal'.
}
\description{
Creates 'auditor_model_halfnormal' object that can be used for plotting halfnormal plot.
}
\examples{
library(DALEX)
data(DALEX::titanic)
titanic <- na.omit(titanic[1:100,])
model_glm <- glm(survived ~ ., family = binomial, data = titanic)
exp_glm <- explain(model_glm)
library(audit)
model_halfnormal(exp_glm)

}
\references{
Moral, R., Hinde, J., & Demétrio, C. (2017). Half-Normal Plots and Overdispersed Models in R: The hnp Package.doi:http://dx.doi.org/10.18637/jss.v081.i10
}
