% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_halfnormal.R
\name{model_halfnormal}
\alias{model_halfnormal}
\title{Create Model Fit explainer}
\usage{
model_halfnormal(object, quant = FALSE, ...)
}
\arguments{
\item{object}{An object of class ModelAudit.}

\item{quant}{if TRUE values on axis are on quantile scale.}

\item{...}{other parameters passed do \code{\link[hnp]{hnp}} function.}
}
\value{
An object of the class 'model_halfnormal'.
}
\description{
Creates modelFit object to be plotted.
}
\examples{
library(DALEX)
data(titanic)
titanic <- na.omit(titanic[1:100,])
model_glm <- glm(survived ~ ., family = binomial, data = titanic)
audit_glm <- audit(model_glm)

model_halfnormal(audit_glm)

}
