% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_residual.R
\name{model_residual}
\alias{model_residual}
\alias{modelResiduals}
\title{Create Model Residuals Explaination}
\usage{
model_residual(x, variable = NULL)

modelResiduals(x, variable = NULL)
}
\arguments{
\item{variable}{Name of model variable to order residuals.
If value is NULL the data is ordered by a vector of actual response (\code{y} parameter
passed to the \code{\link{explain}} function). One can also pass any name of any other variable
in the data set. If \code{variable = ""} is set, unordered observations are presented.}

\item{object}{An object of class 'explainer' created with function \code{\link[explain]{DALEX}} from the DALEX package.}
}
\description{
Creates 'auditor_model_residual' that contains sorted residuals. An object can be further used to generate plots.
For the list of possible plots see see also section.
}
\examples{
library(DALEX)
data(DALEX::titanic)
titanic <- na.omit(titanic)
model_glm <- glm(survived ~ ., family = binomial, data = titanic)
exp_glm <- explain(model_glm, data = titanic, y = titanic$survived)

library(auditor)
model_residual(exp_glm)

}
\seealso{
\code{\link{plot_acf}, \link{plot_autocorrelation}, \link{plot_residual}, \link{plot_residual_boxplot},
\link{plot_pca}, \link{plot_correlation}, \link{plot_prediction}, \link{plot_rec}, \link{plot_residual_density},
\link{plot_residual}, \link{plot_rroc}, \link{plot_scalelocation}, \link{plot_tsecdf}}
}
