% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_residual.R
\name{model_residual}
\alias{model_residual}
\alias{modelResiduals}
\title{Create Model Residuals explainer}
\usage{
model_residual(x, variable = NULL)

modelResiduals(x, variable = NULL)
}
\arguments{
\item{variable}{Name of model variable to order residuals.
If value is NULL the data is ordered by a vector of actual response (\code{y} parameter
passed to the \code{\link{explain}} function). One can also pass any name of any other variable
in the data set. If \code{variable = ""} is set, unordered observations are presented.}

\item{object}{An object of class 'ModelAudit'explain' created with functio \code{\link[explain]{DALEX}} from the DALEX package.}
}
\description{
Creates modelResiduals object to be plotted.
}
\examples{
library(DALEX)
data(titanic)
titanic <- na.omit(titanic)

model_glm <- glm(survived ~ ., family = binomial, data = titanic)

exp_glm <- explain(model_glm, data = titanic, y = titanic$survived)

library(auditor)
model_residual(exp_glm)

}
