% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFit.R
\name{modelFit}
\alias{modelFit}
\title{Create Model Fit explainer}
\usage{
modelFit(object, quant.scale = FALSE, ...)
}
\arguments{
\item{object}{An object of class ModelAudit.}

\item{quant.scale}{if TRUE values on axis are on quantile scale.}

\item{...}{other parameters passed do \code{\link[hnp]{hnp}} function.}
}
\description{
Creates modelFit object to be plotted.
}
\examples{
library(MASS)
model.glm <- glm(Postwt ~ Prewt + Treat + offset(Prewt), family = gaussian, data = anorexia)
audit.glm <- audit(model.glm)

mf.glm <- modelFit(audit.glm)

}
