% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPerformance.R
\name{modelPerformance}
\alias{modelPerformance}
\title{Create Model Performance explainer}
\usage{
modelPerformance(object, variable = NULL)
}
\arguments{
\item{object}{An object of class ModelAudit.}

\item{variable}{Optional. Name of variable to order residuals. If value is NULL data order is taken. If value is "Predicted response" or "Fitted values" then data is ordered by fitted values. If value is "Observed response" the data is ordered by a vector of actual response (\code{y} parameter passed to the \code{\link{audit}} function).}
}
\description{
Creates model Performance object to be plotted.
}
