% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_audit.R,
%   R/plot_model_cooksdistance.R, R/plot_model_evaluation.R,
%   R/plot_model_halfnormal.R, R/plot_model_performance.R,
%   R/plot_model_residual.R
\name{plot.model_audit}
\alias{plot.model_audit}
\alias{plot.model_cooksdistance}
\alias{plot.model_evaluation}
\alias{plot.model_halfnormal}
\alias{plot.auditor_model_performance}
\alias{plot.model_residual}
\title{Model diagnostic plots}
\usage{
\method{plot}{model_audit}(x, ..., type = "residual", ask = TRUE,
  grid = TRUE)

\method{plot}{model_cooksdistance}(x, ..., type = "residual",
  ask = TRUE, grid = TRUE)

\method{plot}{model_evaluation}(x, ..., type = "residual", ask = TRUE,
  grid = TRUE)

\method{plot}{model_halfnormal}(x, ..., type = "residual", ask = TRUE,
  grid = TRUE)

\method{plot}{auditor_model_performance}(x, ..., type = "residual",
  ask = TRUE, grid = TRUE)

\method{plot}{model_residual}(x, ..., type = "residual", ask = TRUE,
  grid = TRUE)
}
\arguments{
\item{x}{object of class 'model_audit', 'model_residual', 'auditor_model_performance', 'model_evaluation',
'model_cooksdistance', or 'model_halfnormal'.}

\item{...}{other arguments dependent on the type of plot or additionl objects of class modelAudit}

\item{type}{the type of plot. Possible values: 'acf', 'autocorrelation', 'cooksdistance', 'halfnormal', 'residuals',
'lift', 'pca', 'radar', 'correlation', 'prediction', 'rec', 'resiual', 'residual_boxplot','residual_density',
'roc', 'rroc', 'scalelocation', 'tsecdf' (for detailed description see functions in see also section).}

\item{ask}{logical; if TRUE, the user is asked before each plot, see \code{\link[graphics]{par}(ask=)}.}

\item{grid}{logical; if TRUE plots will be plotted on the grid.}
}
\description{
This function provides several diagnostic plots for regression and classification models.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)
plot(lm_au)

library(randomForest)
rf_model <- randomForest(life_length~., data = dragons)
rf_au <- audit(rf_model, data = dragons, y = dragons$life_length)
plot(lm_au, rf_au, type = "radar")


}
\seealso{
\code{\link{plot_acf}, \link{plot_autocorrelation}, \link{plot_cooksdistance},
\link{plot_halfnormal}, \link{plot_residual}, \link{plot_residual_boxplot}, \link{plot_lift}, \link{plot_pca},
\link{plot_radar}, \link{plot_correlation},
\link{plot_prediction}, \link{plot_rec}, \link{plot_residual_density}, \link{plot_residual}, \link{plot_roc},
\link{plot_rroc}, \link{plot_scalelocation}, \link{plot_tsecdf}}
}
