% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ModelAudit.R
\name{plot.modelAudit}
\alias{plot.modelAudit}
\title{Model diagnostioc plots}
\usage{
\method{plot}{modelAudit}(x, type = "Residuals", ...)
}
\arguments{
\item{x}{object of class ModelAudit}

\item{type}{the type of score to  be calculated. Possible values: 'ACF', 'Autocorrelation', 'Cook',
'HalfNormal', 'Residuals', 'ScaleLocation' (see the details section).}

\item{...}{other arguments dependent on the type of plot.}
}
\description{
This function provides many scores
}
\details{
TODO: desription of plot types
}
\seealso{
\code{\link{plotACF}, \link{plotAutocorrelation}, \link{plotCook}, \link{plotHalfNormal},
\link{plotResiduals}, \link{plotScaleLocation}}
}
