% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ModelAudit.R
\name{plot.modelAudit}
\alias{plot.modelAudit}
\title{plot}
\usage{
\method{plot}{modelAudit}(x, variable, ...)
}
\arguments{
\item{x}{object of class ModelAudit}

\item{variable}{name of variable to order residuals}

\item{...}{further arguments passed to or from other methods.}
}
\description{
plot
}
