% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ModelAudit.R
\name{plot.modelAudit}
\alias{plot.modelAudit}
\title{Model diagnostioc plots}
\usage{
\method{plot}{modelAudit}(x, ..., type = "Residuals")
}
\arguments{
\item{x}{object of class ModelAudit}

\item{...}{other arguments dependent on the type of plot.}

\item{type}{the type of score to  be calculated. Possible values: 'ACF', 'Autocorrelation', 'Cook',
'HalfNormal', 'Residuals', 'ScaleLocation', 'ROC', 'RROC', 'REC' (see the details section).}
}
\description{
This function provides many scores
}
\details{
TODO: desription of plot types
}
\seealso{
\code{\link{plotACF}, \link{plotAutocorrelation}, \link{plotCook}, \link{plotHalfNormal},
\link{plotResiduals}, \link{plotScaleLocation}}, \link{plotROC}, \link{plotRROC}, \link{plotREC}
}
