% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ModelAudit.R
\name{plot.ModelAudit}
\alias{plot.ModelAudit}
\title{plot}
\usage{
\method{plot}{ModelAudit}(x, ...)
}
\arguments{
\item{x}{object of class ModelAudit}

\item{...}{further arguments passed to or from other methods.}
}
\description{
plot
}
