% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ModelAudit.R, R/plot_modelFit.R,
%   R/plot_modelPerformance.R, R/plot_modelResiduals.R,
%   R/plot_observationInfluence.R
\name{plot.modelAudit}
\alias{plot.modelAudit}
\alias{plot.modelFit}
\alias{plot.modelPerformance}
\alias{plot.modelResiduals}
\alias{plot.observationInfluence}
\title{Model diagnostic plots}
\usage{
\method{plot}{modelAudit}(x, ..., type = "Residual", ask = TRUE,
  grid = TRUE)

\method{plot}{modelFit}(x, ..., type = "Residual", ask = TRUE,
  grid = TRUE)

\method{plot}{modelPerformance}(x, ..., type = "Residual", ask = TRUE,
  grid = TRUE)

\method{plot}{modelResiduals}(x, ..., type = "Residual", ask = TRUE,
  grid = TRUE)

\method{plot}{observationInfluence}(x, ..., type = "Residual",
  ask = TRUE, grid = TRUE)
}
\arguments{
\item{x}{object of class modelAudit, modelResiduals or observationInfluence.}

\item{...}{other arguments dependent on the type of plot or additionl objects of class modelAudit}

\item{type}{the type of plot. Possible values: 'ACF', 'Autocorrelation', 'CooksDistance', 'HalfNormal', 'Residuals', 'LIFT',
ModelPCA', 'ModelRanking', ModelCorrelation', 'Prediction', 'REC', 'Resiual', 'ResidualBoxplot',ResidualDensity', 'ROC', 'RROC',
ScaleLocation', 'TwoSidedECDF' (for detailed description see functions in see also section).}

\item{ask}{logical; if TRUE, the user is asked before each plot, see \code{\link[graphics]{par}(ask=)}.}

\item{grid}{logical; if TRUE plots will be plotted on the grid.}
}
\description{
This function provides several diagnostic plots for regression and classification models.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)
plot(lm_au)

library(randomForest)
rf_model <- randomForest(life_length~., data = dragons)
rf_au <- audit(rf_model, data = dragons, y = dragons$life_length)
plot(lm_au, rf_au, type = "ModelRanking")


}
\seealso{
\code{\link{plotACF}, \link{plotAutocorrelation}, \link{plotCooksDistance},
\link{plotHalfNormal}, \link{plotResidual}, \link{plotResidualBoxplot}, \link{plotLIFT}, \link{plotModelPCA}, \link{plotModelRanking}, \link{plotModelCorrelation},
\link{plotPrediction}, \link{plotREC}, \link{plotResidualDensity}, \link{plotResidual}, \link{plotROC},
\link{plotRROC}, \link{plotScaleLocation}, \link{plotTwoSidedECDF}}
}
