% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modelAudit.R
\name{plot.modelAudit}
\alias{plot.modelAudit}
\title{Model diagnostic plots}
\usage{
plot.modelAudit(x, ..., type = "Residuals")
}
\arguments{
\item{x}{object of class modelAudit}

\item{...}{other arguments dependent on the type of plot or additionam objects of class modelAudit}

\item{type}{the type of plot. Possible values: 'ACF', 'Autocorrelation', 'Cook',
'HalfNormal', 'Residuals', 'ScaleLocation', 'ROC', 'RROC', 'REC' (for detailed description see functions in seealso section).}
}
\description{
This function provides several diagnostic plots for regression and classification models.
}
\seealso{
\code{\link{plotACF}, \link{plotAutocorrelation}, \link{plotCook}, \link{plotHalfNormal},
\link{plotResiduals}, \link{plotScaleLocation}, \link{plotROC}, \link{plotRROC}, \link{plotREC}}
}
