% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_acf.R
\name{plot_acf}
\alias{plot_acf}
\alias{plotACF}
\title{Autocorrelation Function Plot}
\usage{
plot_acf(object, ..., variable = NULL, alpha = 0.95)

plotACF(object, ..., variable = NULL, alpha = 0.95)
}
\arguments{
\item{object}{An object of class 'auditor_model_residual' created with \code{\link{model_residual}} function.}

\item{...}{Other 'auditor_model_residual' objects to be plotted together. Please, note that they should be created with the same value
of parameter 'variable'.}

\item{alpha}{Confidence level of the interval.}
}
\value{
A ggplot object.
}
\description{
Plot Autocorrelation Function of models residuals.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_exp <- DALEX::explain(lm_model, data = dragons, y = dragons$life_length)
library(auditor)
plot_acf(lm_exp)

library(randomForest)
rf_model <- randomForest(life_length~., data = dragons)
rf_exp <- explain(rf_model, data = dragons, y = dragons$life_length)
plot_acf(lm_exp, rf_exp)


}
