% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_autocorrelation.R
\name{plot_autocorrelation}
\alias{plot_autocorrelation}
\alias{plotAutocorrelation}
\title{Autocorrelation of Residuals Plot}
\usage{
plot_autocorrelation(object, ..., variable = "_y_hat_", smooth = FALSE)

plotAutocorrelation(object, ..., variable, smooth = FALSE)
}
\arguments{
\item{object}{An object of class 'auditor_model_residual' created with \code{\link{model_residual}} function.}

\item{...}{Other 'auditor_model_residual' objects to be plotted together.}

\item{variable}{Name of variable to order residuals on a plot.
If \code{variable="_y_"}, the data is ordered by a vector of actual response (\code{y} parameter
passed to the \code{\link[DALEX]{explain}} function).}

\item{smooth}{Logical, if TRUE smooth line will be added.}
}
\value{
A ggplot object.
}
\description{
Plot of i-th residual vs i+1-th residual.
}
\examples{
dragons <- DALEX::dragons[1:100, ]

# fit a model
lm_model <- lm(life_length ~ ., data = dragons)

# use DALEX package to wrap up a model into explainer
lm_exp <- DALEX::explain(lm_model, data = dragons, y = dragons$life_length)

# validate a model with auditor
library(auditor)
lm_mr <- model_residual(lm_exp)

# plot results
plot_autocorrelation(lm_mr)
plot(lm_mr, type = "autocorrelation")
plot_autocorrelation(lm_mr, score = TRUE, smooth = TRUE)
plot(lm_mr, type = "autocorrelation", score = TRUE, smooth = TRUE)

}
