% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation.R
\name{plot_correlation}
\alias{plot_correlation}
\alias{plotModelCorrelation}
\title{Correlation of Model's Residuals Plot}
\usage{
plot_correlation(object, ..., values = "fit")

plotModelCorrelation(object, ..., values = "fit")
}
\arguments{
\item{object}{An object of class 'auditor_model_residual' created with \code{\link{model_residual}} function.}

\item{...}{Other 'auditor_model_residual' objects to be plotted together.}

\item{values}{"fit" for model fitted values or "res" for residual values.}
}
\value{
Invisibly returns a \code{\link[gtable]{gtable}} object.
}
\description{
Matrix of plots. Left-down triangle consists of plots of fitted values (aternatively residuals),
on the diagonal there are density plots of fitted values (alternatively residuals), in the right-top triangle
there are correlations between fitte dvalues (alternatively residuals).
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_exp <- DALEX::explain(lm_model, data = dragons, y = dragons$life_length)
library(auditor)
lm_mr <- model_residual(lm_exp)

library(randomForest)
rf_model <- randomForest(life_length~., data = dragons)
rf_exp <- DALEX::explain(rf_model, data = dragons, y = dragons$life_length)
rf_mr <- model_residual(rf_exp)
plot_correlation(lm_mr, rf_mr)


}
