% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lift.R
\name{plot_lift}
\alias{plot_lift}
\alias{plotLIFT}
\title{LIFT Chart}
\usage{
plot_lift(object, ...)

plotLIFT(object, ...)
}
\arguments{
\item{object}{An object of class 'auditor_model_evaluation' created with \code{\link{model_evaluation}} function.}

\item{...}{Other 'auditor_model_evaluation' objects to be plotted together.}
}
\value{
A ggplot object.docum
}
\description{
LIFT is a plot of the rate of positive prediction against true positive rate for the different thresholds.
It is useful for measuring and comparing the accuracy of the classificators.
}
\examples{
titanic <- na.omit(DALEX::titanic)
titanic$survived <- titanic$survived == "yes"
model_glm <- glm(survived ~ ., family = binomial, data = titanic)
exp_glm <- DALEX::explain(model_glm, data = titanic, y = titanic$survived)
library(auditor)
eva_glm <- model_evaluation(exp_glm)
plot_lift(eva_glm)

model_glm_2 <- glm(survived ~ .-age, family = binomial, data = titanic)
exp_glm_2 <- DALEX::explain(model_glm_2, data = titanic, y = titanic$survived, label = "glm2")
eva_glm_2 <- model_evaluation(exp_glm_2)

plot_lift(eva_glm, eva_glm_2)


}
\seealso{
\code{\link{model_evaluation}}
}
