% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\alias{plotModelPCA}
\title{Principal Component Analysis of models}
\usage{
plot_pca(object, ..., scale = TRUE)

plotModelPCA(object, ..., scale = TRUE)
}
\arguments{
\item{object}{An object of class 'auditor_model_residual' created with \code{\link{model_residual}} function.}

\item{...}{Other 'auditor_model_residual' objects to be plotted together.}

\item{scale}{A logical value indicating whether the models residuals should be scaled before the analysis.}
}
\value{
A ggplot object.
}
\description{
Principal Component Analysis of models residuals.
PCA can be used to assess the similarity of the models.
}
\examples{
dragons <- DALEX::dragons[1:100, ]

# fit a model
lm_model <- lm(life_length ~ ., data = dragons)

# use DALEX package to wrap up a model into explainer
lm_exp <- DALEX::explain(lm_model, data = dragons, y = dragons$life_length)

# validate a model with auditor
library(auditor)
lm_mr <- model_residual(lm_exp)

library(randomForest)
rf_model <- randomForest(life_length~., data = dragons)
rf_exp <- DALEX::explain(rf_model, data = dragons, y = dragons$life_length)
rf_mr <- model_residual(rf_exp)

# plot results
plot_pca(lm_mr, rf_mr)

}
