% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prediction.R
\name{plot_prediction}
\alias{plot_prediction}
\alias{plotPrediction}
\title{Predicted response vs Observed or Variable Values}
\usage{
plot_prediction(object, ..., variable = NULL, smooth = FALSE,
  abline = FALSE)

plotPrediction(object, ..., variable = NULL, smooth = FALSE,
  abline = FALSE)
}
\arguments{
\item{object}{An object of class 'model_audit' or 'model_residual'.}

\item{...}{Other modelAudit or modelResiduals objects to be plotted together.}

\item{variable}{Only for 'model_audit' objects. Name of model variable to order residuals.
If value is NULL the data is ordered by a vector of actual response (\code{y} parameter
passed to the \code{\link{audit}} function). One can also pass any name of any other variable
in the data set. If \code{variable = ""} is set, unordered observations are presented.}

\item{smooth}{Logical, indicates whenever smooth line should be added.}

\item{abline}{Logical, indicates whenever function y = x shoul be added. Works only
with \code{variable = NULL} which is a default option.}
}
\description{
Plot of predicted response vs observed or variable Values.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)
plotPrediction(lm_au, variable = "height", abline = TRUE)

library(randomForest)
rf_model <- randomForest(life_length~., data = dragons)
rf_au <- audit(rf_model, data = dragons, y = dragons$life_length)
plotPrediction(lm_au, rf_au, variable = "height", smooth = TRUE)

}
\seealso{
\code{\link{plot.model_audit}}
}
