% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prediction.R
\name{plot_prediction}
\alias{plot_prediction}
\alias{plotPrediction}
\title{Predicted response vs Observed or Variable Values}
\usage{
plot_prediction(object, ..., variable = "_y_", smooth = FALSE,
  abline = FALSE)

plotPrediction(object, ..., variable = NULL, smooth = FALSE,
  abline = FALSE)
}
\arguments{
\item{object}{An object of class 'auditor_model_residual.}

\item{...}{Other modelAudit or modelResiduals objects to be plotted together.}

\item{variable}{Name of variable to order residuals on a plot.
If \code{variable="_y_"}, the data is ordered by a vector of actual response (\code{y} parameter
passed to the \code{\link[DALEX]{explain}} function).
If \code{variable = "_y_hat_"} the data on the plot will be ordered by predicted response.
If \code{variable = NULL}, unordered observations are presented.}

\item{smooth}{Logical, indicates whenever smooth line should be added.}

\item{abline}{Logical, indicates whenever function y = x shoul be added. Works only
with \code{variable = NULL} which is a default option.}
}
\value{
A ggplot2 object.
}
\description{
Plot of predicted response vs observed or variable Values.
}
\examples{
dragons <- DALEX::dragons[1:100, ]

# fit a model
lm_model <- lm(life_length ~ ., data = dragons)

 # use DALEX package to wrap up a model into explainer
lm_exp <- DALEX::explain(lm_model, data = dragons, y = dragons$life_length)

# validate a model with auditor
library(auditor)
lm_mr <- model_residual(lm_exp)

# plot results
plot_prediction(lm_mr)
plot_prediction(lm_mr, variable = "height", abline = TRUE)
plot(lm_mr, type = "prediction", abline = TRUE)

library(randomForest)
rf_model <- randomForest(life_length~., data = dragons)
rf_exp <- DALEX::explain(rf_model, data = dragons, y = dragons$life_length)
rf_mr <- model_residual(rf_exp)
plot_prediction(lm_mr, rf_mr, variable = "height", smooth = TRUE)


}
