% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_radar.R
\name{plot_radar}
\alias{plot_radar}
\alias{plotModelRanking}
\title{Model Ranking Plot}
\usage{
plot_radar(object, ..., score = c("mae", "mse", "rec", "rroc"),
  new_score = NULL, print = TRUE)

plotModelRanking(object, ..., score = c("MAE", "MSE", "REC", "RROC"),
  new_score = NULL)
}
\arguments{
\item{object}{An object of class ModelAudit.}

\item{...}{Other model_audit objects to be plotted together.}

\item{score}{Vector of score names to be plotted.}

\item{new_score}{A named list of functions that take one argument: object of class ModelAudit and return a numeric value. The measure calculated by the function should have the property that lower score value indicates better model.}

\item{print}{logical, should values of scores be printed?}
}
\value{
ggplot object
}
\description{
Radar plot with model score. score are scaled to [0,1], each score is inversed and divided by maximum score value.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)
library(randomForest)
rf_model <- randomForest(life_length~., data = dragons)
rf_au <- audit(rf_model, data = dragons, y = dragons$life_length)
plot_radar(lm_au, rf_au)

}
\seealso{
\code{\link{plot.model_audit}}
}
