% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residual_boxplot.R,
%   R/plot_residual_density.R
\name{plot_residual_boxplot}
\alias{plot_residual_boxplot}
\alias{plotResidualBoxplot}
\alias{plotResidualDensity}
\title{Plot Boxplots of Residuals}
\usage{
plot_residual_boxplot(object, ...)

plotResidualBoxplot(object, ...)

plotResidualDensity(object, ..., split = FALSE, variable = NULL)
}
\arguments{
\item{object}{An object of class 'model_audit'.}

\item{...}{Other modelAudit objects to be plotted together.

dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)
plotResidual(lm_au)

library(randomForest)
rf_model <- randomForest(life_length~., data = dragons)
rf_au <- audit(rf_model, data = dragons, y = dragons$life_length)
plotResidualBoxplot(lm_au, rf_au)}
}
\description{
A plot of residuals.
}
\seealso{
\code{\link{plot.model_audit}}
}
