% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residual_density.R
\name{plot_residual_density}
\alias{plot_residual_density}
\title{Residual Density Plot}
\usage{
plot_residual_density(object, ..., split = FALSE, variable = NULL)
}
\arguments{
\item{object}{An object of class 'model_audit'.}

\item{...}{Other modelAudit objects to be plotted together.}

\item{split}{Logical. Indicates whenever plot should be splitted by variable.}

\item{variable}{variable name o split. Optional. Should be provided  only for modelAudit object.}
}
\value{
ggplot object
}
\description{
Density of model residuals.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_exp <- DALEX::explain(lm_model, data = dragons, y = dragons$life_length)
lm_mr <- model_residual(lm_exp)
plot_residual_density(lm_mr)

library(randomForest)
rf_model <- randomForest(life_length~., data = dragons)
rf_exp <- DALEX::explain(rf_model, data = dragons, y = dragons$life_length)
rf_mr <- model_residual(rf_exp)
plot_residual_density(lm_mr, rf_mr)

}
\seealso{
\code{\link{plot.model_audit}}
}
