% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotACF.R
\name{plotACF}
\alias{plotACF}
\title{ACF plot}
\usage{
plotACF(object, variable)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{variable}{name of variable to order residuals}
}
\description{
Plot autocorrelation function of model residuals.
Runs test, Durbin-Watson test
}
