% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAutocorrelation.R
\name{plotAutocorrelation}
\alias{plotAutocorrelation}
\title{Autocorrelation plot}
\usage{
plotAutocorrelation(object, variable = NULL, score = TRUE)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{variable}{"Fitted values" or name of dependent or independent variable to order residuals. If NULL residuals won't be ordered..}

\item{score}{Logical, if TRUE values of \link{scoreDW} and \link{scoreRuns} will be added}
}
\description{
Plot i-th residual vs i+1-th residual. Checking autocorrelation of residuals.
Runs score, Durbin-Watson score
}
