% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAutocorrelation.R
\name{plotAutocorrelation}
\alias{plotAutocorrelation}
\title{Autocorrelation plot}
\usage{
plotAutocorrelation(object, variable = NULL)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{variable}{name of dependent or independent variable to order residuals. If NULL the fitted values are taken.}
}
\description{
Plot i-th residual vs i+1-th residual. Checking autocorrelation of residuals.
Runs score, Durbin-Watson score
}
