% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotD3_autocorrelation.R
\name{plotD3_autocorrelation}
\alias{plotD3_autocorrelation}
\alias{plotD3Autocorrelation}
\title{Autocorrelation Plot in D3 with r2d3 package.}
\usage{
plotD3_autocorrelation(object, ..., variable = NULL, points = TRUE,
  smooth = FALSE, point_count = NULL, single_plot = TRUE,
  scale_plot = FALSE, background = FALSE)

plotD3Autocorrelation(object, ..., variable = NULL, points = TRUE,
  smooth = FALSE, point_count = NULL, single_plot = TRUE,
  scale_plot = FALSE, background = FALSE)
}
\arguments{
\item{object}{An object of class modelAudit or modelResiduals.}

\item{...}{Other modelAudit or modelResiduals objects to be plotted together.}

\item{variable}{Only for modelAudit objects. Name of model variable to order residuals.
If value is NULL the data is ordered by a vector of actual response (\code{y} parameter
passed to the \code{\link{audit}} function). One can also pass any name of any other variable
in the data set. If \code{variable = ""} is set, unordered observations are presented.}

\item{points}{Logical, indicates whenever observations should be added as points. By defaul it's TRUE.}

\item{smooth}{Logical, indicates whenever smoothed lines should be added. By default it's FALSE.}

\item{point_count}{Number of points to be plotted per model. Points will be chosen randomly. By default plot all of them.}

\item{single_plot}{Logical, indicates whenever single or facets should be plotted. By default it's TRUE.}

\item{scale_plot}{Logical, indicates whenever the plot should scale with height. By default it's FALSE.}

\item{background}{Logical, available only if single_plot = FALSE. Indicates whenever backgroud plots should be plotted. By default it's FALSE.}
}
\value{
a `r2d3` object.
}
\description{
Plot of i-th residual vs i+1-th residual.
}
\examples{
library(auditor)
library(DALEX)

lm_model <- lm(m2.price ~ ., data = apartments)

library("randomForest")
rf_model <- randomForest(m2.price ~ ., data = apartments)

lm_au <- audit(lm_model, label = "lm", data = apartments, y = apartments$m2.price)
rf_au <- audit(rf_model, label = "rf", data = apartments, y = apartments$m2.price)

lm_mr_year <- modelResiduals(lm_au, variable = "construction.year")
rf_mr_year <- modelResiduals(rf_au, variable = "construction.year")

plotD3_autocorrelation(rf_mr_year, smooth = FALSE, scale_plot = TRUE, point_count = 1000)

plotD3_autocorrelation(rf_mr_year, lm_mr_year, smooth = TRUE)
plotD3_autocorrelation(rf_mr_year, lm_mr_year, smooth = TRUE, single_plot = FALSE)

}
