% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotD3ACF.R
\name{plotD3ACF}
\alias{plotD3ACF}
\title{Plot Autocorrelation Function in D3 with r2d3 package.}
\usage{
plotD3ACF(object, ..., variable = NULL, alpha = 0.95,
  scale_plot = FALSE)
}
\arguments{
\item{object}{An object of class modelAudit or modelResiduals.}

\item{...}{Other modelAudit or modelResiduals objects to be plotted together.}

\item{variable}{Only for modelAudit object. Name of model variable to order residuals. If value is NULL data order is taken. If value is "Predicted response" or "Fitted values" then data is ordered by fitted values. If value is "Observed response" the data is ordered by a vector of actual response (\code{y} parameter passed to the \code{\link{audit}} function).}

\item{alpha}{Confidence level of the interval.}

\item{scale_plot}{Logical, indicates whenever the plot should scale with height. By default it's FALSE.}
}
\value{
a `r2d3` object.
}
\description{
Plot Autocorrelation Function of models residuals.
}
\examples{
library(car)
lm_model <- lm(prestige~education + women + income, data = Prestige)
lm_au <- audit(lm_model, data = Prestige, y = Prestige$prestige)
plotD3ACF(lm_au)

library(randomForest)
rf_model <- randomForest(prestige~education + women + income, data = Prestige)
rf_au <- audit(rf_model, data = Prestige, y = Prestige$prestige)
plotD3ACF(lm_au, rf_au)

}
