% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotD3_lift.R
\name{plotD3LIFT}
\alias{plotD3LIFT}
\title{Plot LIFT in D3 with r2d3 package.}
\usage{
plotD3LIFT(object, ..., scale_plot = FALSE)
}
\arguments{
\item{object}{An object of class modelAudit or modelEvaluation.}

\item{...}{Other modelAudit objects to be plotted together.}

\item{scale_plot}{Logical, indicates whenever the plot should scale with height. By default it's FALSE.}
}
\value{
a `r2d3` object
}
\description{
LIFT is a plot of the rate of positive prediction against true positive rate for the different thresholds.
It is useful for measuring and comparing the accuracy of the classificators.
}
\examples{
library(DALEX)
data(titanic)
titanic <- na.omit(titanic)
titanic$survived <- titanic$survived == "yes"
model_glm <- glm(survived ~ ., family = binomial, data = titanic)
audit_glm <- audit(model_glm, data = titanic, y = titanic$survived)

plotD3LIFT(audit_glm)

model_glm_2 <- glm(survived ~ .-age, family = binomial, data = titanic)
audit_glm_2 <- audit(model_glm_2, data = titanic, y = titanic$survived, label = "glm2")

plotD3LIFT(audit_glm, audit_glm_2, scale_plot = TRUE)

}
\seealso{
\code{\link{plotLIFT}}
}
