% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotD3REC.R
\name{plotD3REC}
\alias{plotD3REC}
\title{Regression Error Characteristic Curves (REC) in D3 with r2d3 Package.}
\usage{
plotD3REC(object, ..., scale_plot = FALSE)
}
\arguments{
\item{object}{An object of class modelAudit or modelResiduals.}

\item{...}{Other modelAudit or model Residuals objects to be plotted together.}

\item{scale_plot}{Logical, indicates whenever the plot should scale with height. By default it's FALSE.}
}
\value{
r2d3 object
}
\description{
Error Characteristic curves are a generalization of ROC curves.
On the x axis of the plot there is an error tolerance and on the y axis there is a percentage
of observations predicted within the given tolerance.
}
\details{
REC curve estimates the Cumulative Distribution Function (CDF) of the error

Area Over the REC Curve (REC) is a biased estimate of the expected error
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)
plotD3REC(lm_au)

library(randomForest)
rf_model <- randomForest(life_length~., data = dragons)
rf_au <- audit(rf_model, data = dragons, y = dragons$life_length)
plotD3REC(lm_au, rf_au)

}
\references{
Bi J., Bennett K.P. (2003). Regression error characteristic curves, in: Twentieth
International Conference on Machine Learning (ICML-2003), Washington, DC.
}
\seealso{
\code{\link{plot.modelAudit}, \link{plotROC}, \link{plotRROC}, \link{plotREC}}
}
