% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotD3Residual.R
\name{plotD3Residual}
\alias{plotD3Residual}
\title{Plot Residuals vs Observed, Fitted or Variable Values in D3 with r2d3 Package.

Function \code{plotD3Residual} plots resudial values vs observed, fitted or variable values in the model.
It uses output from \code{modelAudit} or \code{modelResiduals} function.}
\usage{
plotD3Residual(object, ..., variable = NULL, points = TRUE,
  smooth = FALSE, std_residuals = FALSE, point_count = NULL,
  single_plot = TRUE, scale_plot = FALSE, background = FALSE)
}
\arguments{
\item{object}{An object of class modelAudit or modelResiduals.}

\item{...}{Other modelAudit or modelResiduals objects to be plotted together.}

\item{variable}{Name of model variable to order residuals. If value is NULL data order is taken or variable from modelResiduals object. If value is "Predicted response" or "Fitted values" then data is ordered by fitted values. If value is "Observed response" the data is ordered by a vector of actual response (\code{y} parameter passed to the \code{\link{audit}} function).}

\item{points}{Logical, indicates whenever observations should be added as points. By defaul it's TRUE.}

\item{smooth}{Logical, indicates whenever smoothed lines should be added. By default it's FALSE.}

\item{std_residuals}{Logical, indicates whenever standardized residuals should be used. By default it's FALSE.}

\item{point_count}{Number of points to be plotted per model. Points will be chosen randomly. By default plot all of them.}

\item{single_plot}{Logical, indicates whenever single or facets should be plotted. By default it's TRUE.}

\item{scale_plot}{Logical, indicates whenever the plot should scale with height. By default it's FALSE.}

\item{background}{Logical, available only if single_plot = FALSE. Indicates whenever backgroud plots should be plotted. By default it's FALSE.}
}
\value{
an `r2d3` object.
}
\description{
Plot Residuals vs Observed, Fitted or Variable Values in D3 with r2d3 Package.

Function \code{plotD3Residual} plots resudial values vs observed, fitted or variable values in the model.
It uses output from \code{modelAudit} or \code{modelResiduals} function.
}
\examples{
library(auditor)
library(car)

lm_model <- lm(prestige~education + women + income, data = Prestige)
lm_au <- audit(lm_model, label = "lm")
plotD3Residual(lm_au, variable = "income")

library(randomForest)
rf_model <- randomForest(prestige~education + women + income, data = Prestige)
rf_au <- audit(rf_model, label = "rf")
rf_mr <- modelResiduals(rf_au, "income")
plotD3Residual(lm_au, rf_mr, variable = "income", smooth = TRUE)
plotD3Residual(lm_au, rf_mr, variable = "income", smooth = TRUE, single_plot = F)

}
\seealso{
\code{\link{plotResidual}}
}
