% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHalfNormal.R
\name{plotHalfNormal}
\alias{plotHalfNormal}
\title{Half-Normal plot}
\usage{
plotHalfNormal(object, xlab = "Theoretical quantiles", ylab = "residuals",
  ...)
}
\arguments{
\item{object}{fitted model object or numeric vector}

\item{xlab}{the text for the x axis}

\item{ylab}{the text for the y axis}

\item{...}{extra arguments passed to \link[hnp]{hnp}.}
}
\value{
An object of class ggplot
}
\description{
Function \code{plotHalfNormal}...
}
