% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHalfNormal.R
\name{plotHalfNormal}
\alias{plotHalfNormal}
\title{Half-Normal plot}
\usage{
plotHalfNormal(object, score = TRUE, quant.scale = FALSE,
  xlab = "Half-Normal Quantiles", ylab = "Residuals", main = "", ...)
}
\arguments{
\item{object}{ModelAudit object, fitted model object or numeric vector}

\item{score}{if TRUE score based on probability density function}

\item{quant.scale}{if TRUE values on avis are on quantile scale}

\item{xlab}{the text for the x axis}

\item{ylab}{the text for the y axis}

\item{main}{title of plot}

\item{...}{extra arguments passed to \link[hnp]{hnp}.}
}
\value{
An object of class ggplot
}
\description{
The half-normal plot is one of the tools designed to evaluate the goodness of fit of a
statistical models. It is a graphical method for comparing two probability distributions by plotting
their quantiles against each other.
Points on the plot correspond to ordered absolute values of model diagnostic
(i.e. standardized residuals) plotted against theoretical order statistics from a half-normal distribution.
}
\seealso{
\code{\link{scoreHalfNormal}}
}
