% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLift.R
\name{plotLIFT}
\alias{plotLIFT}
\title{LIFT}
\usage{
plotLIFT(object, ...)
}
\arguments{
\item{object}{An object of class modelAudit or modelEvaluation.}

\item{...}{Other modelAudit objects to be plotted together.}
}
\value{
ggplot object
}
\description{
LIFT is a plot of the rate of positive prediction against true positive rate for the different thresholds.
It is useful for measuring and comparing the accuracy of the classificators.
}
\examples{
library(DALEX)
data(titanic)
titanic <- na.omit(titanic)
titanic$survived <- titanic$survived == "yes"
model_glm <- glm(survived ~ ., family = binomial, data = titanic)
audit_glm <- audit(model_glm, data = titanic, y = titanic$survived)

plotLIFT(audit_glm)

model_glm_2 <- glm(survived ~ .-age, family = binomial, data = titanic)
audit_glm_2 <- audit(model_glm_2, data = titanic, y = titanic$survived, label = "glm2")
meva_glm_2 <- modelEvaluation(audit_glm_2)

plotLIFT(audit_glm, audit_glm_2)


}
\seealso{
\code{\link{plot.modelAudit}}
}
