% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelCorrelation.R
\name{plotModelCorrelation}
\alias{plotModelCorrelation}
\title{Model Correlation Plot}
\usage{
plotModelCorrelation(object, ..., values = "fitted")
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{...}{other modelAudit objects to be plotted together}

\item{values}{"fitted" for model fitted values or "residuals" for residuals values}
}
\value{
ggplot object
}
\description{
Matrix of plots
}
\seealso{
\code{\link{plot.modelAudit}}
}
