% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelCorrelation.R
\name{plotModelCorrelation}
\alias{plotModelCorrelation}
\title{Model Correlation Plot}
\usage{
plotModelCorrelation(object, ..., values = "Fitted values")
}
\arguments{
\item{object}{An object of class ModelAudit.}

\item{...}{Other modelAudit objects to be plotted together.}

\item{values}{"Fitted values" or "Predicted response" for model fitted values or "Residuals" for residual values.}
}
\value{
ggplot object
}
\description{
Matrix of plots
}
\seealso{
\code{\link{plot.modelAudit}}
}
