% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelPCA.R
\name{plotModelPCA}
\alias{plotModelPCA}
\title{Principal Component Analysis of models}
\usage{
plotModelPCA(object, ..., scale = TRUE)
}
\arguments{
\item{object}{An object of class modelAudit or modelResiduals,}

\item{...}{Other modelAudit or modelResiduals objects to be plotted together.}

\item{scale}{A logical value indicating whether the models residuals should be scaled before the analysis.}
}
\value{
ggplot object
}
\description{
Principal Component Analysis of models residuals.
PCA can be used to assess the similarity of the models.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)

library(randomForest)
rf_model <- randomForest(life_length~., data = dragons)
rf_au <- audit(rf_model, data = dragons, y = dragons$life_length)

plotModelPCA(lm_au, rf_au)

}
\seealso{
\code{\link{plot.modelAudit}}
}
