% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelPCA.R
\name{plotModelPCA}
\alias{plotModelPCA}
\title{Principal Component Analysis of models}
\usage{
plotModelPCA(object, ..., scale = TRUE, invisible = "none")
}
\arguments{
\item{object}{An object of class ModelAudit,}

\item{...}{Other modelAudit objects to be plotted together.}

\item{scale}{A logical value indicating whether the models residuals should be scaled bfore the analysis.}

\item{invisible}{A text specifying the elements to be hidden on the plot. Default value is "none". Allowed values are "model", "observ".}
}
\value{
ggplot object
}
\description{
Principal Component Analysis of models residuals.
PCA can be used to assess the similarity of the models.
}
\seealso{
\code{\link{plot.modelAudit}}
}
