% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelRanking.R
\name{plotModelRanking}
\alias{plotModelRanking}
\title{Model Ranking Plot}
\usage{
plotModelRanking(object, ..., type = c("MAE", "MSE", "REC", "RROC"),
  new.score = NULL)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{...}{other modelAudit objects to be plotted together}

\item{type}{vector of score names to be plotted.}

\item{new.score}{a function or named list of functions that take one argument: object of class ModelAudit and return a numeric value.}
}
\value{
ggplot object
}
\description{
Radar plot with model scores. Scores are scaled to [0,1], each score is inversed and divided by maximum score value.
}
\seealso{
\code{\link{plot.modelAudit}}
}
