% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPredict.R
\name{plotPrediction}
\alias{plotPrediction}
\title{Observed response values vs Predicted response}
\usage{
plotPrediction(object, ..., variable = NULL)
}
\arguments{
\item{object}{An object of class modelAudit}

\item{...}{other modelAudit objects to be plotted together}
}
\description{
Plot of predicted vs actual response.
}
\seealso{
\code{\link{plot.modelAudit}}
}
