% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResidDens.R
\name{plotResidDens}
\alias{plotResidDens}
\title{Residual Density Plot}
\usage{
plotResidDens(object, ...)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{...}{other modelAudit objects to be plotted together}
}
\value{
ggplot object
}
\description{
Density of model residuals.
}
\seealso{
\code{\link{plot.modelAudit}}
}
