% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResidual.R
\name{plotResidual}
\alias{plotResidual}
\title{Plot Residuals vs Fitted or Variable Values}
\usage{
plotResidual(object, ..., variable = NULL)
}
\arguments{
\item{object}{An object of class modelAudit}

\item{...}{other modelAudit objects to be plotted together}

\item{variable}{name of modle variable for x-axis. If NULL fitted values are taken.}
}
\description{
A plot of residuals against fitted values or any variable values.
}
\seealso{
\code{\link{plot.modelAudit}}
}
