% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResidualDensity.R
\name{plotResidualDensity}
\alias{plotResidualDensity}
\title{Residual Density Plot}
\usage{
plotResidualDensity(object, ..., split = FALSE, variable = NULL)
}
\arguments{
\item{object}{An object of class ModelAudit.}

\item{...}{Other modelAudit objects to be plotted together.}

\item{split}{Logical. Indicates whenever plot should be splitted by variable.}

\item{variable}{variable name o split. Optional. Should be provided  only for modelAudit object.}
}
\value{
ggplot object
}
\description{
Density of model residuals.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)
plotResidualDensity(lm_au)

library(randomForest)
rf_model <- randomForest(life_length~., data = dragons)
rf_au <- audit(rf_model, data = dragons, y = dragons$life_length)
plotResidualDensity(lm_au, rf_au)

}
\seealso{
\code{\link{plot.modelAudit}}
}
