% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResiduals.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Plot Residuals vs Fitted or Variable Values}
\usage{
plotResiduals(object, variable = NULL)
}
\arguments{
\item{object}{An object of class modelAudit}

\item{variable}{name of modle variable for x-axis. If NULL fitted values are taken.}
}
\description{
A plot of residuals against fitted values or any variable values.
}
\seealso{
\code{\link{plot.modelAudit}}
}
