% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotROC.R, R/plotShapMap.R
\name{plotROC}
\alias{plotROC}
\alias{plotROC}
\title{Receiver Operating Characteristic (ROC)}
\usage{
plotROC(object, ...)

plotROC(object, ...)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{...}{other modelAudit objects to be plotted together}

\item{object}{An object of class ModelAudit}

\item{...}{other modelAudit objects to be plotted together}
}
\value{
ggplot object

ggplot object
}
\description{
Receiver Operating Characterstic Curve is a plot of the true positive rate (TPR) against the false positive rate (FPR) for the different thresholds.
It is useful for measuring and comparing the accuracy of the classificators.
}
\examples{
library(auditor)
library(mlbench)
data("PimaIndiansDiabetes")

model.glm <- glm(diabetes~., family=binomial,	data=PimaIndiansDiabetes)
au.glm <- audit(model.glm, label="class glm")
plotROC(au.glm)

model.glm.press <- glm(diabetes~pressure, family=binomial,	data=PimaIndiansDiabetes)
au.glm.press <- audit(model.glm.press)
plotROC(au.glm, au.glm.press)

}
\seealso{
\code{\link{plot.modelAudit}, \link{plotRROC}, \link{plotREC}}

\code{\link{plot.modelAudit}}
}
