% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScaleLocation.R
\name{plotScaleLocation}
\alias{plotScaleLocation}
\title{Scale location plot}
\usage{
plotScaleLocation(object, variable = NULL)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{variable}{name of dependent or independent variable to order residuals. If NULL the fitted values are taken.}
}
\description{
Variable values vs square root of the absolute value of the residuals.
A vertical line corresponds to median.
Goldfeld-Quandt score - checking assumption of homoscedasticity of residuals.
}
