% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTwoSidedECDF.R
\name{plotTwoSidedECDF}
\alias{plotTwoSidedECDF}
\title{Two-sided Cumulative Distribution Function}
\usage{
plotTwoSidedECDF(object, ..., error.scaled = TRUE)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{...}{other modelAudit objects to be plotted together}

\item{error.scale}{Should ECDF be scaled by proportions of positive and negative proportions?}
}
\value{
ggplot object
}
\description{
Two-sided Cumulative Distribution Function
}
\seealso{
\code{\link{plot.modelAudit}}
}
