% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tsecdf.R
\name{plotTwoSidedECDF}
\alias{plotTwoSidedECDF}
\title{Two-sided Cumulative Distribution Function}
\usage{
plotTwoSidedECDF(object, ..., error.scaled = TRUE, outliers = NA,
  residuals = TRUE, y.reversed = FALSE)
}
\arguments{
\item{object}{An object of class ModelAudit.}

\item{...}{Other modelAudit objects to be plotted together.}

\item{error.scaled}{A logical value indicating whether ECDF should be scaled by proportions of positive and negative proportions.}

\item{outliers}{Number of outliers to be marked.}

\item{residuals}{A logical value indicating whether residuals should be marked.}

\item{y.reversed}{A logical value indicating whether values on y axis should be reversed.}
}
\value{
ggplot object
}
\description{
Cumulative Distribution Function for positive and negative residuals.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)
plotTwoSidedECDF(lm_au)

library(randomForest)
rf_model <- randomForest(life_length~., data = dragons)
rf_au <- audit(rf_model, data = dragons, y = dragons$life_length)
plotTwoSidedECDF(lm_au, rf_au, y.reversed = TRUE)

}
\seealso{
\code{\link{plot.model_audit}}
}
