% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTwoSidedECDF.R
\name{plotTwoSidedECDF}
\alias{plotTwoSidedECDF}
\title{Two-sided Cumulative Distribution Function}
\usage{
plotTwoSidedECDF(object, ..., error.scaled = TRUE, outliers = NA,
  residuals = TRUE, y.reversed = FALSE)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{...}{other modelAudit objects to be plotted together}

\item{error.scaled}{Should ECDF be scaled by proportions of positive and negative proportions?}

\item{outliers}{number of outliers to be marked}

\item{residuals}{should residuals be marked?}

\item{y.reversed}{should values on y axib be reversed}
}
\value{
ggplot object
}
\description{
Two-sided Cumulative Distribution Function
}
\seealso{
\code{\link{plot.modelAudit}}
}
