% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_support_function.R
\name{prepare_object}
\alias{prepare_object}
\title{Prepare object for `make_dataframe`` function}
\usage{
prepare_object(object, variable, nlabel, type, quant)
}
\arguments{
\item{object}{An audited model}

\item{variable}{Variable}

\item{nlabel}{Number of labels}

\item{type}{Type of model passed}

\item{quant}{Logical}
}
\description{
Prepare object for `make_dataframe`` function
}
