% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Model scores computations}
\usage{
score(object, score = "GQ", ...)
}
\arguments{
\item{object}{object An object of class ModelAudit}

\item{score}{the type of score to  be calculated. Possible values: 'GQ', 'DW', 'Runs', 'HalfNormal', 'Cook' (see the details section).}

\item{...}{other arguments dependent on the type of score.}
}
\description{
This function provides many scores
}
\details{
TODO: desription of scores
}
\seealso{
\code{\link{scoreGQ}, \link{scoreDW}, \link{scoreRuns}, \link{scoreHalfNormal}, \link{scoreCook}}
}
