% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_auc.R
\name{score_auc}
\alias{score_auc}
\alias{scoreROC}
\title{Area Under ROC Curve (AUC)}
\usage{
score_auc(object)

scoreROC(object)
}
\arguments{
\item{object}{An object of class 'explainer'.}
}
\value{
an object of class scoreAudit
}
\description{
Area Under Curve (AUC) for Receiver Operating Characteristic.
}
\examples{
library(DALEX)
data(titanic)
titanic <- na.omit(titanic)
titanic$survived <- titanic$survived == "yes"
model_glm <- glm(survived ~ ., family = binomial, data = titanic)
exp_glm <- DALEX::explain(model_glm, y = titanic$survived)

score_auc(exp_glm)

}
\seealso{
\code{\link{plot_roc}}
}
