% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_mae.R
\name{score_mae}
\alias{score_mae}
\alias{scoreMAE}
\title{Mean Absolute Error}
\usage{
score_mae(object)

scoreMAE(object)
}
\arguments{
\item{object}{An object of class 'explainer' created with function \code{\link[DALEX]{explain}} from the DALEX package.}
}
\value{
an object of class 'score_audit'
}
\description{
Mean Absolute Error.
}
\examples{
dragons <- DALEX::dragons[1:100, ]

# fit a model
model_lm <- lm(life_length ~ ., data = dragons)

# create an explainer
exp_lm <- DALEX::explain(model_lm, data = dragons, y = dragons$life_length)

# calculate score
score_mae(exp_lm)


}
\seealso{
\code{\link{score}}
}
