% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_rmse.R
\name{score_rmse}
\alias{score_rmse}
\alias{scoreRMSE}
\title{Root Mean Square Error}
\usage{
score_rmse(object)

scoreRMSE(object)
}
\arguments{
\item{object}{An object of class ModelAudit.}
}
\value{
an object of class scoreAudit
}
\description{
Root Mean Square Error.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)
score_rmse(lm_au)


}
\seealso{
\code{\link{score}}
}
