% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_rroc.R
\name{score_rroc}
\alias{score_rroc}
\title{Area Over the Curve for RROC Curves}
\usage{
score_rroc(object)
}
\arguments{
\item{object}{An object of class 'model_audit'.}
}
\value{
an object of class 'score_audit'.
}
\description{
The area over the Regression Receiver Operating Characteristic.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)
score_rroc(lm_au)


}
\references{
Hernández-Orallo, José. 2013. ‘ROC Curves for Regression’. Pattern Recognition 46 (12): 3395–3411.
}
\seealso{
\code{\link{plot_rroc}}
}
