% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreCook.R
\name{scoreCook}
\alias{scoreCook}
\title{Score based on Cooks distance}
\usage{
scoreCook(object, print = TRUE)
}
\arguments{
\item{object}{An object of class ModelAudit}

\item{print}{If TRUE progress is printed}
}
\description{
Cooks Distance TODO: defifnition here
}
