% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreDW.R
\name{scoreDW}
\alias{scoreDW}
\title{Durbin-Watson Score}
\usage{
scoreDW(object, variable = NULL)
}
\arguments{
\item{object}{object An object of class ModelAudit}

\item{variable}{"Fitted values" or name of dependent or independent variable to order residuals. If NULL residuals won't be ordered..}
}
\description{
Autocorrelation based on Durbin-Watson Test.
}
