% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreGQ.R
\name{scoreGQ}
\alias{scoreGQ}
\title{Goldfeld-Quandt Score}
\usage{
scoreGQ(object, variable = NULL)
}
\arguments{
\item{object}{object An object of class ModelAudit}

\item{variable}{name of dependent or independent variable to order residuals. If NULL the fitted values are taken.}
}
\description{
This score is calculated on the basis of Goldfeld-Quandt test, which is used for checking for homoscedasticity of residuals in regression analyses.

Goldfeld-Quandt test is based on a comparison of the residual sum of squares (RSS) using the F-statistic.
The test consists of dividing dataset into two groups. Then estimating models separately for each subset and calculating
the residual sum of squares for each group (RSS1 and RSS2).

The test statistic is the ratio of the mean square residual errors for two groups, which corresponds to the F-test of equality of variances.
\eqn{F = (MSE_1)/(MSE_2)}
where \eqn{MSE = (RSS)/(n-p)}
where n is the number of observations and p is the number of variables .

The score value is helpful in comparing models. It is worth ponting out that results of tests like p-value makes sense only
when the test assumptions are satisfied. Otherwise test statistic may be considered as a score.
\code{scoreGQ} function uses a two-sided F-test.
}
