% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMAE.R
\name{scoreMAE}
\alias{scoreMAE}
\title{Mean Absolute Error}
\usage{
scoreMAE(object)
}
\arguments{
\item{object}{An object of class ModelAudit or modelResiduals.}
}
\value{
an object of class scoreAudit
}
\description{
Mean Absolute Error.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)
scoreMAE(lm_au)


}
\seealso{
\code{\link{score}}
}
