% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMSE.R
\name{scoreMSE}
\alias{scoreMSE}
\title{Mean Square Error}
\usage{
scoreMSE(object)
}
\arguments{
\item{object}{An object of class ModelAudit.}
}
\value{
an object of class scoreAudit
}
\description{
Mean Square Error.
}
\seealso{
\code{\link{score}}
}
