% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_mse.R
\name{scoreMSE}
\alias{scoreMSE}
\title{Mean Square Error}
\usage{
scoreMSE(object)
}
\arguments{
\item{object}{An object of class modelAudit or modelResiduals.}
}
\value{
an object of class scoreAudit
}
\description{
Mean Square Error.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)
scoreMSE(lm_au)

}
\seealso{
\code{\link{score}}
}
