% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreROC.R
\name{scoreROC}
\alias{scoreROC}
\title{Area Under Curve (AUC)}
\usage{
scoreROC(object, type = "AUC")
}
\arguments{
\item{object}{An object of class ModelAudit}
}
\value{
numeric
}
\description{
Area Under Curve (AUC) for Receiver Operating Characteristic
}
\seealso{
\code{\link{plotROC}}
}
