% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreROC.R
\name{scoreROC}
\alias{scoreROC}
\title{Area Under ROC Curve (AUC)}
\usage{
scoreROC(object)
}
\arguments{
\item{object}{An object of class ModelAudit.}
}
\value{
an object of class scoreAudit
}
\description{
Area Under Curve (AUC) for Receiver Operating Characteristic.
}
\examples{
library(DALEX)
data(titanic)
titanic <- na.omit(titanic)
titanic$survived <- titanic$survived == "yes"
model_glm <- glm(survived ~ ., family = binomial, data = titanic)
audit_glm <- audit(model_glm, y = titanic$survived)

scoreROC(audit_glm)

}
\seealso{
\code{\link{plotROC}}
}
