% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreRuns.R
\name{scoreRuns}
\alias{scoreRuns}
\title{Runs Score}
\usage{
scoreRuns(object, variable = NULL)
}
\arguments{
\item{object}{object An object of class ModelAudit}

\item{variable}{name of dependent or independent variable to order residuals. If NULL the fitted values are taken.}
}
\description{
Score based on Runs test statistic. Note that this test is not very strong. It utilizes only signs of the residuals.
}
