% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_runs.R
\name{scoreRuns}
\alias{scoreRuns}
\title{Runs Score}
\usage{
scoreRuns(object, variable = NULL)
}
\arguments{
\item{object}{object An object of class ModelAudit.}

\item{variable}{name of model variable to order residuals. If value is NULL data order is taken. If value is "Predicted response" or "Fitted values" then data is ordered by fitted values. If value is "Observed response" the data is ordered by a vector of actual response (\code{y} parameter passed to the \code{\link{audit}} function).}
}
\description{
Score based on Runs test statistic. Note that this test is not very strong. It utilizes only signs of the residuals.
The score value is helpful in comparing models. It is worth pointing out that results of tests like p-value makes sense only
when the test assumptions are satisfied. Otherwise test statistic may be considered as a score.
}
\examples{
dragons <- DALEX::dragons[1:100, ]
lm_model <- lm(life_length ~ ., data = dragons)
lm_au <- audit(lm_model, data = dragons, y = dragons$life_length)
scoreRuns(lm_au)

}
