% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDW.R
\name{testDW}
\alias{testDW}
\title{Durbin-Watson Test}
\usage{
testDW(model, vars = NULL)
}
\arguments{
\item{model}{model}

\item{vars}{name of variable - the observations in the model are ordered by the values of this variable. If NULL - first variable is taken.}
}
\description{
Test for autocorrelation with the Durbin-Watson Test.
}
