% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGQ.R
\name{testGQ}
\alias{testGQ}
\title{Goldfeld-Quandt Test}
\usage{
testGQ(model, vars = NULL)
}
\arguments{
\item{model}{model}

\item{vars}{name of variable - the observations in the model are ordered by the values of this variable. If NULL - first variable is taken.}
}
\description{
Test for homooskedasticity with the Goldfeld-Quandt Test.
}
